% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{annotateMetaClusters}
\alias{annotateMetaClusters}
\title{Add cluster labels to the metaClusters column of a survival data.frame}
\usage{
annotateMetaClusters(metaClusterSurvival, clusterLabels)
}
\arguments{
\item{metaClusterSurvival}{A \code{data.frame} containing per sample
metaclusters and survival data, as returned by the `mergeSurvOn` function
in this package.}

\item{clusterLabels}{A \code{character} vector with labels for each metacluster
in `metaClusterSurvival`. Matches by postion, with metacluster 1 being labelled
with the first item in `clusterLabels` and so on.}
}
\value{
The `metaClusterSurvival` \code{data.frame} with the 'metaClusters'
  column labelled.
}
\description{
Add cluster labels to the metaClusters column of a survival data.frame
}
