% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSignatureComparison.R
\name{annotateSampleMetaClassDT}
\alias{annotateSampleMetaClassDT}
\title{Swap numeric metaclusters to names in the 'metaClasses' column}
\usage{
annotateSampleMetaClassDT(sampleMetaClassDT, clusterLabels)
}
\arguments{
\item{sampleMetaClassDT}{A \code{data.table} containing per cohort per sample
meta-class labels.}

\item{clusterLabels}{A \code{character} vector of meta-class labels matched
according to vector index against the 'metaClasses' column.}
}
\value{
A \code{data.table} with 'metaClasses' as a character column labelled
   with `clusterLabels`.
}
\description{
Swap numeric metaclusters to names in the 'metaClasses' column
}
