% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{annotateSampleMetaClasses}
\alias{annotateSampleMetaClasses}
\title{Replace the metacluster indexes with name of the metaclusters}
\usage{
annotateSampleMetaClasses(allConClusters, clusterwiseSamples)
}
\arguments{
\item{allConClusters}{A \code{list} of consensus clustering results, as
returned byt he `conClustAllCohorts` function.}

\item{clusterwiseSamples}{A \code{data.table} with the samples per cluster
in a list column, as returne dby the `getClusterwiseSamples` function.}
}
\description{
Replace the metacluster indexes with name of the metaclusters
}
