% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assignColDataColumn}
\alias{assignColDataColumn}
\title{assignColDataColumn

Assign a new column directly to the colData slot of a SummarizedExperiment.}
\usage{
assignColDataColumn(SE, colname, values)
}
\arguments{
\item{SE}{A \code{SummarizedExperiment} object to assign colData columns to}

\item{colname}{The name of the column to assign \code{values} to.}

\item{values}{The values to assign to the col}
}
\value{
The \code{SE} object with the new column in the objects \code{colData} slot.
}
\description{
assignColDataColumn

Assign a new column directly to the colData slot of a SummarizedExperiment.
}
\examples{
SE <- SummarizedExperiment(matrix(rnorm(100), 10, 10))
updatedSE <- assignColDataColumn(SE, 'test', rep(1, 10))

}
