% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_aucMetaEstimates.R
\name{aucMetaEstimates}
\alias{aucMetaEstimates}
\title{Return}
\usage{
aucMetaEstimates(validationCohorts, validationStats, seqCohorts)
}
\arguments{
\item{validationCohorts}{A}

\item{validationStats}{A}
}
\value{
A named \code{list} of \code{list}s containing AUC, standard error of AUC,
and the assocaited p-value for each cohort as well as summaries for
sequencing cohorts, microarray cohorts as well as an overall summary.
The list is named the same as validationCohorts.
}
\description{
Return
}
