% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_aucMetaEstimates.R
\name{AUCstats}
\alias{AUCstats}
\title{Calculate AUC and its standard error for a list of validation cohorts}
\usage{
AUCstats(formattedValCohorts, probList)
}
\arguments{
\item{formattedValCohorts}{A \code{list} of validation cohorts.
formatted using `formatValidationCohorts`.}

\item{probList}{A \code{list} of per cohort survival probability prediction.}
}
\description{
Calculate AUC and its standard error for a list of validation cohorts
}
