% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-barPlotModelComparison.R
\name{barPlotModelComparison}
\alias{barPlotModelComparison}
\title{Make A Bar Plot Comparing Perforamnce Between Two \code{S4} Obects Representing
Mathematical Models.}
\usage{
barPlotModelComparison(model1, model2, ...)
}
\arguments{
\item{model1}{An\code{S4} object containing results of a mathematical model}

\item{model2}{An \code{S4} object containing results of a different mathematical
model, but with the same or overalapping samples.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
A bar plot comparing some aspect of model1 and model2
}
\description{
Make A Bar Plot Comparing Perforamnce Between Two \code{S4} Obects Representing
Mathematical Models.
}
\examples{
data(sampleCohortList)
data(sampleTrainedPCOSPmodel)
data(sampleICGCmicro)

# Setup the models
clinicalModel <- ClinicalModel(sampleICGCmicro,
  formula='prognosis ~ sex + age + T + N + M + grade')

# Train the models
trainedClinicalModel <- trainModel(clinicalModel)

# Make predctions
PCOSPpredCohortList <- predictClasses(sampleCohortList[c('PCSI', 'TCGA')],
  model=sampleTrainedPCOSPmodel)
clinicalPredCohortList <- predictClasses(sampleCohortList[c('PCSI', 'TCGA')],
  model=trainedClinicalModel)

# Validate the models
validatedPCOSPmodel <- validateModel(sampleTrainedPCOSPmodel,
  valData=PCOSPpredCohortList)
validatedClinicalModel <- validateModel(trainedClinicalModel,
  valData=clinicalPredCohortList)

# Plot the comparison
modelCompBarPlot <- barPlotModelComparison(validatedClinicalModel,
 validatedPCOSPmodel, stat='AUC')

}
