% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{birnbaum}
\alias{birnbaum}
\title{Published classifier gene signature for Birnbaum}
\description{
The genes and coefficients for the gene signature from
Birnbaum \emph{et al.} (2017)
}
\examples{
# Loads chen, birnbaum and haiderSigScores objects
data(existingClassifierData)

}
