% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerCellineDrugSensitivities.R
\name{boxplotAUCperSubtypePerDataset}
\alias{boxplotAUCperSubtypePerDataset}
\title{Boxplot the distribution of AUC values for each subtype in each dataset}
\usage{
boxplotAUCperSubtypePerDataset(mergedDT, conIdx, palette = "Set1")
}
\arguments{
\item{mergedDT}{A \code{data.table}}

\item{conIdx}{A \code{numeric}}

\item{paletee}{A \code{character} vector specifying the RColorBrewer palette
to use for the plot. Passed to `ggpubr::ggboxplot`s `palette` argument.}
}
\value{
A \code{list} object
}
\description{
Boxplot the distribution of AUC values for each subtype in each dataset
}
