% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerComparison.R
\name{boxplotBiomarkerScores}
\alias{boxplotBiomarkerScores}
\title{Boxplot the normalized per sample gene expression for the specified gene,
    grouped by meta-class.}
\usage{
boxplotBiomarkerScores(
  geneBiomarkerScores,
  comparisons,
  geneName,
  palette = "Set1",
  saveDir,
  fileName
)
}
\arguments{
\item{geneBiomarkerScores}{A \code{data.table} containing per sample expression
values for the gene specified in `geneName`. As returned by the
`computeGeneBiomarkerScores` function from this package.}

\item{comparisons}{A \code{list} of length two `character` vectors containing
the meta-classes to calculate p-value for difference in means between.}

\item{geneName}{A \code{character} vector with the name of the gene being
boxplotted. Used to label the y-axis}

\item{palette}{A \code{character} vector specifying the name of a palette
from RColorBrewer for the plot. Passed as `palette` argument to
`ggpubr:ggboxplot`. Defaults to 'Set1'.}

\item{saveDir}{An optional \code{character} vector specifying the path
to the directory where the plot should be saved. If excluded, fileName
will not work.}

\item{fileName}{An optional \code{character} vector specifying the
name and extension of the file to save the plot it. This is passed to
the `ggplot2::ggsave`.}
}
\value{
A \code{ggplot} object of the resulting boxplot.
}
\description{
Boxplot the normalized per sample gene expression for the specified gene,
    grouped by meta-class.
}
