% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerComparison.R
\name{boxplotBiomarkerScoresL}
\alias{boxplotBiomarkerScoresL}
\title{Create a list of ggplots comparing the meta-class expression for each biomarker}
\usage{
boxplotBiomarkerScoresL(
  geneBiomarkerScoreL,
  biomarkers,
  comparisons,
  palette = "Set1"
)
}
\arguments{
\item{biomarkers}{A \code{character} vector of biomarker gene names, with
each gene named for the meta-cluster it is a biomarker in.}

\item{palette}{A \code{character} vector specifying the name of a palette
from RColorBrewer for the plot. Passed as `palette` argument to
`ggpubr:ggboxplot`. Defaults to 'Set1'.}

\item{geneBiomarkerscoreL}{A \code{list} of \code{data.table}s containing
per sample expression values for the gene specified in `geneName`.}

\item{comparison}{A \code{list} of character vectors containing the names
of cohorts to compare. All comparisons must be pairwise (i.e., each
character vector can have only two names).}
}
\value{
A \code{list} of `ggplot`as
}
\description{
Create a list of ggplots comparing the meta-class expression for each biomarker
}
