% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellularityComparison.R
\name{boxplotCellarityByCohort}
\alias{boxplotCellarityByCohort}
\title{Compare the cellularity between meta}
\usage{
boxplotCellarityByCohort(mergedDT, saveDir, fileName)
}
\arguments{
\item{mergedDT}{A \code{data.table} of per sample cellularity and
predicted meta-classes for several cohorts.}

\item{saveDir}{An optional \code{character} vector specifying the path
to the directory where the plot should be saved. If excluded, fileName
will not work.}

\item{fileName}{An optional \code{character} vector specifying the
name and extension of the file to save the plot it. This is passed to
the `ggplot2::ggsave`.}
}
\value{
A \code{ggplot} object with a grouped boxplot comparing the
   distribution of sample cellularity between the predicted meta-classes.
}
\description{
Compare the cellularity between meta
}
