% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compassSubtypeWaterfall.R
\name{boxplotClassNLR}
\alias{boxplotClassNLR}
\title{Boxplot difference between in neutrophil-lymphocyte ratio between
  meta-classes, comparin means with `method`}
\usage{
boxplotClassNLR(
  classSurvCompDT,
  method = "kruskal.test",
  palette = "Set1",
  saveDir,
  fileName
)
}
\arguments{
\item{classSurvCompDT}{A \code{data.table} containing meta-class labels and
per sample survival data.}

\item{method}{A \code{character} vector specifying the mean comparison method
passed to `ggpubr::stat_compare_means`.}

\item{saveDir}{An optional \code{character} vector specifying the path
to the directory where the plot should be saved. If excluded, fileName
will not work.}

\item{fileName}{An optional \code{character} vector specifying the
name and extension of the file to save the plot it. This is passed to
the `ggplot2::ggsave`.}
}
\value{
A \code{gpplot} with predicted class on the x-axis and the
   neutrophil-lymphcyte ratio on the y-axis.
}
\description{
Boxplot difference between in neutrophil-lymphocyte ratio between
  meta-classes, comparin means with `method`
}
