% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDXdrugSubtyping.R
\name{boxplotPDXsubtypePerDrug}
\alias{boxplotPDXsubtypePerDrug}
\title{Boxplot per drug AAC values between subclasses}
\usage{
boxplotPDXsubtypePerDrug(PDXmergedDT)
}
\arguments{
\item{PDXmergedDT}{A \code{data.table} of merge patient derived xenograph
drug sensitivity and predicted meta-class/subtype.}
}
\value{
A \code{list} of ggplot objects, one for each drug in `PDXmergedDT`.
   This can be format to a plot grid using the `ggarrangePlotL` function.
}
\description{
Boxplot per drug AAC values between subclasses
}
