% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_buildPCOSPmodel.R
\name{buildPCOSPmodels}
\alias{buildPCOSPmodels}
\title{Build PCOSP Model from input data}
\usage{
buildPCOSPmodels(trainingCohorts, numModels, nthread, saveDir)
}
\arguments{
\item{trainingCohorts}{A named \code{list} of training cohorts for which
to fit and select PCOSP models.}

\item{nthread}{\code{integer} The number of threads to parallelize across}

\item{saveDir}{\code{character} A path to a directory to save the model. If you
exclude this the function will return the model object instead.}
}
\value{
\code{list} Either returns the model object or, is \code{saveDir} is
  specified it saves to disk instead and return the path
}
\description{
The script is used for building Pancreatic cancer overall survival predictor
  using unique 89 samples profiled using microarray and sequencing platform.
}
\section{Warning}{
 This function uses random numbers; remember to
  \code{set.seed()} before running to ensure reproducible results
}

\examples{
# To ensure reproducible results
set.seed(1987)

# Load the data
data(trainingCohorts)

# Return the object
selectedModels <- buildPCOSPmodels(trainingCohorts, numModels=10, nthread=1)

# OR Save the object to disk
buildPCOSPmodel(traingCohort, numModels=10, saveDir=tempdir())

}
