% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_buildRandomGeneAssignmentModel.R
\name{buildRandomGeneAssignmentModels}
\alias{buildRandomGeneAssignmentModels}
\title{Build PCOSP models with the group labels randomly shuffled}
\usage{
buildRandomGeneAssignmentModels(
  trainingCohorts,
  seqCohort,
  numModels,
  nthread,
  saveDir,
  original
)
}
\arguments{
\item{trainingCohorts}{A named \code{list} of training cohorts for which
to fit and select PCOSP models.}

\item{numModels}{A code{numeric} vector with an integer number of models
to run.}

\item{nthread}{\code{integer} The number of threads to parallelize across}

\item{saveDir}{\code{character} A path to a directory to save the model. If you
exclude this the function will return the model object instead.}

\item{original}{A \code{logical} vector, if true this function calls
the deprecated (and slow) function used in the original PCOSP paper.
This is inlcuded to ensure the paper is reproducible using this pacakge.}
}
\value{
\code{list} Either returns the model object or, is \code{saveDir} is
  specified it saves to disk instead and return the path
}
\description{
Build PCOSP models with the group labels randomly shuffled
}
\section{Warning}{
 This function uses random numbers; remember to
  \code{set.seed()} before running to ensure reproducible results
}

