% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_buildRandomLabelShufflingModel.R
\name{buildRandomLabelShufflingModel}
\alias{buildRandomLabelShufflingModel}
\title{Build PCOSP models with the group labels randomly shuffled}
\usage{
buildRandomLabelShufflingModel(trainingCohorts, numModels, nthread, saveDir)
}
\arguments{
\item{trainingCohorts}{A named \code{list} of training cohorts for which
to fit and select PCOSP models.}

\item{numModels}{A \code{numeric} vector containing an integer number of
models to fit}

\item{nthread}{\code{integer} The number of threads to parallelize across}

\item{filePath}{A \code{character} vector wtih teh path to a directory
to save the model. If you exclude this the function will return the
model object instead.}

\item{fileNames}{A \code{character} vector with the file name to save under.
If excluded this function returns the object instead.}
}
\value{
\code{list} Either returns the model object or, is \code{saveDir} is
  specified it saves to disk instead and return the path
}
\description{
Build PCOSP models with the group labels randomly shuffled
}
\section{Warning}{
 This function uses random numbers; remember to
  \code{set.seed()} before running to ensure reproducible results
}

