% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{calcAllClusterRepro}
\alias{calcAllClusterRepro}
\title{Permutations test for null distribution of of each consensus clustering results}
\usage{
calcAllClusterRepro(
  MSMthresholds,
  allConClusters,
  allProcCohorts,
  reps,
  nthread,
  seed = NULL
)
}
\arguments{
\item{MSMthresholds}{A \code{data.table} where columns 1 through 5 represent}

\item{allConClusters}{A \code{list} of consensus clustering results for
all non-self pair-wise comparisons of cohorts, as returned by the
`conClustAllCohorts` function.}

\item{allProcCohorts}{A \code{list} of preprocessed gene expression data
for each cohort. As returned by the `preprocCohorts` function.}

\item{reps}{A \code{numeric} vector containing the integer number of
random samples to reproduce clustering over.}

\item{nthread}{A \code{numeric} vector containing the integer number of
threads to parallelize over.}

\item{seed}{A \code{numeric} vector containing the desired seed to be used
for sampling.}
}
\description{
Calculate reproducibility statistics for each consensus clustering result
   using random samples from the the second cohort in a comparison.
}
