% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSignatureComparison.R
\name{calcAllGeneSetSigScores}
\alias{calcAllGeneSetSigScores}
\title{Calculate the signature scores for a set of gene signatures}
\usage{
calcAllGeneSetSigScores(geneSigL, cohortsDataL, sampleMetaClassDT)
}
\arguments{
\item{geneSigL}{A \code{list} of gene signatures (character vector of gene
names).}

\item{cohortsDataL}{A \code{list} of gene by sample cohort gene expression
matrixes.}

\item{sampleMetaClassDT}{A \code{data.table} containing the per cohort per
sample meta-class predictions.}
}
\value{
A \code{list} of gene signature score `data.table`s with each item
    returned from the `computeSigScoreDT` function from this package.
}
\description{
Calculate the signature scores for a set of gene signatures
}
