% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{calcAllMSMthresholds}
\alias{calcAllMSMthresholds}
\title{Calculate the MSM thresholds for each non-self pair-wise comparison of
   cohort consensus clustering results.}
\usage{
calcAllMSMthresholds(cohortPairs, allConClusters, allProcCohorts, nthread)
}
\arguments{
\item{nthread}{}
}
\description{
Calculate the MSM thresholds for each non-self pair-wise comparison of
   cohort consensus clustering results.
}
