% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedClassifiers.R
\name{calcAssocStats}
\alias{calcAssocStats}
\title{Calculate the correlation betwen}
\usage{
calcAssocStats(pubClassifSubtypeDT)
}
\arguments{
\item{pubClassifSubtypeDT}{\code{data.table} with the predicted subtype for
each classifier in each cell-line.}
}
\value{
A \code{data.table} with the cramers V and p.value for each pair-wise
  classifier comparison.
}
\description{
Calculate the correlation betwen
}
