% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{calcClusterRepro}
\alias{calcClusterRepro}
\title{Calculate reproduction statistics for consensus clustering between two
   cohorts using `reps` random samples.}
\usage{
calcClusterRepro(conClusters, procCohort, reps = 100, seed = NULL)
}
\arguments{
\item{conClusters}{A \code{list} of concensus clustering results, as returned
by the `conClustCohort` function or as a list item from `conClustAllCohorts`.}

\item{procCohort}{A \code{matrix} of normalized gene expression data with genes
as rows and samples as columns.}

\item{reps}{The number of permutations to use in `clusterRepro` function. Defaults
to 100.}

\item{seed}{The integer seed to set when reproducing clusters.}
}
\value{
A \code{data.table} containing the p.value, Number, Actual.IGP,
  Actual.Size values for each cluster pair-wise cluster comparison.
}
\description{
Calculate reproduction statistics for consensus clustering between two
   cohorts using `reps` random samples.
}
