% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMetaEffectSizeCalculation.R
\name{calcClustMetaEstStats}
\alias{calcClustMetaEstStats}
\title{Calculate meta-estimate statistics for each gene in each sample in each
   cohort}
\usage{
calcClustMetaEstStats(gwCohortsDT)
}
\arguments{
\item{gwCohortsDT}{A long \code{data.table} of gene expression values per
gene per samplw per cohort, annotated with the predicted meta-class/subtype
for each sample in each cohort. As returned by the `makeGenewiseCohortsDT`
function in this package.}
}
\value{
A gene by meta-class \code{data.table} containing the weigthed mean
    of the meta-effectsize (calculated with `effsize::cohen.d`) for each
    predicted subtype across all cohorts and samples.
}
\description{
Calculate meta-estimate statistics for each gene in each sample in each
   cohort
}
