% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_compareClinicalModels.R
\name{calcCohortProbabilties}
\alias{calcCohortProbabilties}
\title{Calcualte two different prediction probabilties under a given model for each
  cohort}
\usage{
calcCohortProbabilties(fitModels, clinicalCohorts)
}
\arguments{
\item{fitModels}{A \code{list} of clinical models, as returned by
`summarizeClinicalModels`}

\item{clinicalCohorts}{A \code{list} of clinical cohorts to calculate
probabilties under the given models}
}
\value{
A \code{list} of
}
\description{
Calcualte two different prediction probabilties under a given model for each
  cohort
}
