% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{calcGenewiseWeightedMADdf}
\alias{calcGenewiseWeightedMADdf}
\title{Calculate the weighted MAD across all cohorts per gene and
    return a `data.frame` of the results.}
\usage{
calcGenewiseWeightedMADdf(cohortMADrankings)
}
\arguments{
\item{cohortMADrankings}{A \code{list} of \code{data.frames} containing
the genes, madValuesa nd ranking for each cohort. As returned by
the `rankAllCohortGenesByMAD` function.`}
}
\value{
A \code{data.frame} with columns genes, weightedMADs and rankings
    ordered by decreasing weightedMAD values.
}
\description{
Calculate the weighted MAD across all cohorts per gene and
    return a `data.frame` of the results.
}
