% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{calcGenewiseWeightedMADs}
\alias{calcGenewiseWeightedMADs}
\title{Calculate the weighted MAD across all cohorts per gene}
\usage{
calcGenewiseWeightedMADs(cohortMADrankings)
}
\arguments{
\item{cohortMADrankings}{A \code{list} of \code{data.frames} containing
the genes, madValuesa nd ranking for each cohort. As returned by
the `rowMADdfs` function.}
}
\value{
A \code{numeric} vector
}
\description{
Calculate the weighted MAD across all cohorts per gene
}
