% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRASdep.R
\name{calcLScore}
\alias{calcLScore}
\title{KRASdep}
\usage{
calcLScore(dat)
}
\arguments{
\item{dat}{\code{vector} A single sample gene expression profile
with gene symbols in names}
}
\value{
L-score
}
\description{
Calculate L-score
}
\examples{
lScore <- calcLScore(dat)

}
