% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_utilities.R
\name{calcMedAbsDev}
\alias{calcMedAbsDev}
\title{Parallelize calculation of the row-wise median absolute deviance}
\usage{
calcMedAbsDev(dataset)
}
\arguments{
\item{dataset}{}
}
\description{
Parallelize calculation of the row-wise median absolute deviance
}
