% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{calcMSMthresholds}
\alias{calcMSMthresholds}
\title{Calculate the MSM threshold between}
\usage{
calcMSMthresholds(cohortPair, allConClusters, allProcCohorts)
}
\arguments{
\item{allProcCohorts}{}
}
\value{
A \code{data.table}
}
\description{
Calculate the MSM threshold between
}
