% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{calcPerGeneWeigthedMAD}
\alias{calcPerGeneWeigthedMAD}
\title{Calculate aggregate MAD per gene across all cohorts, weighted by the sample size of each cohort}
\usage{
calcPerGeneWeigthedMAD(cohortMADrankings)
}
\arguments{
\item{cohortMADrankings}{A \code{list} of \code{data.frames} containing
the genes, madValues and ranking for each cohort. As returned by
the `rankAllCohortGenesByMAD` function.`}
}
\value{
A \code{data.frame} with columns genes, weightedMADs and rankings
    ordered by ranking.
}
\description{
Calculate aggregate MAD per gene across all cohorts, weighted by the sample size of each cohort
}
