% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calcRowMADdf}
\alias{calcRowMADdf}
\title{Parallelized calculation of the row-wise median absolute deviance using `data.table``}
\usage{
calcRowMADdf(dataset)
}
\arguments{
\item{dataset}{A \code{matrix} of cohort gene expression data.
Genes are rows, samples are columns.}
}
\value{
A \code{data.frame} containing the gene names, row-wise median
    absolute deviances and MAD rankings for the specified cohort. The
    \code{data.frame} has an attribute holding the number of columns
    in the original `dataset` for use in other functions.
}
\description{
Parallelized calculation of the row-wise median absolute deviance using `data.table``
}
