% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRASdep.R
\name{calcSScore}
\alias{calcSScore}
\title{Calculate S-score}
\usage{
calcSScore(dat)
}
\arguments{
\item{dat}{\code{vector} A single sample gene expression profile
with gene symbols in names}
}
\value{
S-score
}
\description{
Calculate S-score
}
\examples{
sScore <- calcSScore(dat)

}
