% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSampleClassifier.R
\name{calcTopGenes1Gt2Matrix}
\alias{calcTopGenes1Gt2Matrix}
\title{Calculate top two genes for each biomarker}
\usage{
calcTopGenes1Gt2Matrix(cohortTopGenesMat, topGenes1, topGenes2, genePairs)
}
\arguments{
\item{cohortTopGenesMat}{A \code{matrix} of normalized gene expression
values for the genes in `topGenes1` and `topGenes2`. Subset from the
matrix returned by `computeCohortCommMat` function in this pacakge.}

\item{topGenes1}{A \code{character} vector of genes names for the top scoring
pairs item predicted to have higher expression on average.}

\item{topGenes2}{A \code{character} vector of gene names for the top scoring
gene pairs item predicted to have lower expression on avareage.}

\item{genePairs}{A \code{character} vector with `topGenes1` and `topGenes`
pasted together with '>'. This is used for label the gene pairs and can
be any character vector of length equal to `topGenes`.`}
}
\value{
A gene-pair by sample binary \code{matrix} where 1 represents gene1
    is greater than gene2 (accurate prediction) and 0 represents gene2 less
    than gene1 (an inaccurate prediction).
}
\description{
Calculate top two genes for each biomarker
}
