% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calculateCohortSubtypeStats.R
\name{calculateCohortSubtypeStats}
\alias{calculateCohortSubtypeStats}
\title{Return a list of lists of classification data.frames, top level is sequencing, microarray, overall;
    subsequent level is both, basal, classical}
\usage{
calculateCohortSubtypeStats(
  probList,
  cohortData,
  cohortClasses,
  seqCohorts,
  hetero
)
}
\arguments{
\item{probList}{A \code{list} of per cohort survival probability predictions}

\item{cohortData}{A \code{list} of per cohort survival data}

\item{cohortClasses}{A \code{list} of per cohort subtype classifications.}

\item{seqCohorts}{A \code{character} vector of names for the cohorts
containing sequencing data. Other cohorts are assumed to contain
microarray data.}

\item{hetero}{A \code{logical} vector passed to `combine.est` for the
combined, sequencing and microarray cohorts, respectively.}
}
\description{
Return a list of lists of classification data.frames, top level is sequencing, microarray, overall;
    subsequent level is both, basal, classical
}
