% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerCellineDrugSensitivities.R
\name{calculateMetaStats}
\alias{calculateMetaStats}
\title{Calculate meta concordance index and meta p-values from a list of per
  dataset concordance index results.}
\usage{
calculateMetaStats(concordanceInds)
}
\arguments{
\item{concordanceInds}{A \code{list} of per dataset concordance index
results for a specfic meta-class.}
}
\value{
A \code{data.table} with the columns 'drugs', 'metaCI', and 'metaPval'
}
\description{
Calculate meta concordance index and meta p-values from a list of per
  dataset concordance index results.
}
