% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calcualteClinicalModelDandCindex.R
\name{calculateModelDandCindex}
\alias{calculateModelDandCindex}
\title{Calculate the D and concordance index for a a list of model predictions and
    clinical cohorts}
\usage{
calculateModelDandCindex(
  modelProbabilities,
  clinicalFeatures,
  seqCohorts,
  model = 1
)
}
\arguments{
\item{modelProbabilities}{A \code{list} of per cohort survival probability
predictions.}

\item{clinicalFeatures}{A \code{list} of per cohort expression data.}

\item{seqCohorts}{A \code{character} vector with the names of the
cohorts containing sequencing data. Others are assumed to be
from microarray data.}

\item{model}{A \code{glm} for predicting survival probabilities.}
}
\description{
Calculate the D and concordance index for a a list of model predictions and
    clinical cohorts
}
