% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calculateValidationStats.R
\name{calculatePCOSPscores}
\alias{calculatePCOSPscores}
\title{Calculate the cohort-wise PCOSP score from a list of validation cohorts}
\usage{
calculatePCOSPscores(validationCohorts, selectedModels, nthread)
}
\arguments{
\item{validationCohorts}{A \code{list} of validation cohorts}

\item{selectedModels}{A \code{list} of selected models, as returned by the
`buildPCOSPmodels` function.}

\item{nthread}{A \code{numeric} vector containing the integer number of threads
to a parallelize across.}
}
\value{
A \code{list} of PCOSP scores for each validation cohorot
}
\description{
Calculate the cohort-wise PCOSP score from a list of validation cohorts
}
