% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calculateValidationStats.R
\name{calculateValidationStats}
\alias{calculateValidationStats}
\title{Validate the Meta Estimate Calculation}
\usage{
calculateValidationStats(
  validationCohorts,
  selectedModels,
  seqCohorts,
  nthread,
  saveDir
)
}
\arguments{
\item{validationCohorts}{A named \code{list} of validation cohorts.}

\item{selectedModels}{A \code{list} of selected models from the
`buildPCOSPmodels` function.}

\item{seqCohorts}{A \code{character} vector specifying which cohorts
are from sequencing data. It is assumed that non-sequencing cohorts
are from micro-array cohorts.}

\item{nthread}{A \code{numeric} vector with an integer number of thread to
parallelize across.}

\item{saveDir}{A \code{character} vector specifying the directory in which
to save results. If missing the function returns the data.}
}
\value{
A named \code{list} of statistical meta-estimates for the combined,
  sequencing and microarray cohorts supplied in validationCohorts. Top level
  list labels indicate which cohorts, with the subsequent list level
  indicating the type of statistic (i.e., dindex vs concordance index)
}
\description{
The function calculates D-index and Concordance index for independent cohorts
  and calculates the meta-estimate of C-index and D-index. The forestplot of
  C-index and D-index was plotted for all the cohorts.
}
