% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-censorSurvival.R
\name{censorSurvival}
\alias{censorSurvival}
\title{Remove Deaths Due to Disease Severity from An S4 Object Where Columns Are
  Samples.}
\usage{
censorSurvival(object, ...)
}
\arguments{
\item{object}{An `S4` object containing survival data to which needs to be
censored.}
}
\value{
`S4` The object subset to only those patients which pass the censoring
  criteria.
}
\description{
Remove Deaths Due to Disease Severity from An S4 Object Where Columns Are
  Samples.
}
