% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ClinicalModel.R
\docType{class}
\name{ClinicalModel-class}
\alias{ClinicalModel-class}
\alias{.ClinicalModel}
\title{Constructor for a SurvivalModel Object.}
\value{
A `SurvivalModel` object.
}
\description{
Constructor for a SurvivalModel Object.
}
