% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CohortList.R
\name{CohortList}
\alias{CohortList}
\title{Constructor for the \code{CohortList} class, a specialized list for stored two or
more \code{SurvivalExperiment} objects.}
\usage{
CohortList(...)
}
\arguments{
\item{...}{One or more \code{SurvivalExperiment} objects.}
}
\description{
Constructor for the \code{CohortList} class, a specialized list for stored two or
more \code{SurvivalExperiment} objects.
}
