% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_compareClinicalModels.R
\name{compareClinicalModels}
\alias{compareClinicalModels}
\title{Compare clinical models between patient cohorts}
\usage{
compareClinicalModels(
  clinicalFeatures,
  cohortClasses,
  cohorts,
  models,
 
    formula = "binary_grp ~ Age + Sex +\\n                                  T_status + N + M + Grade"
)
}
\arguments{
\item{clinicalFeatures}{A}

\item{cohortClasses}{A}

\item{cohorts}{A}

\item{models}{A \code{character} vector names for cohorts in `clinicalFeatures`
to fit a linear model for. All models will be compared against each
cohort in `cohorts`, or all cohorts if `cohorts` is not specified.}

\item{formula}{A}
}
\value{
A \code{list} with the first level representing the model the
  data was fit to, the second level representing the cohort the model was
  compared to, and the third level the associated statistics for that
  comparison.
}
\description{
Compare clinical models between patient cohorts
}
