% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{compareClusters}
\alias{compareClusters}
\title{Compare the consensus clustering results for cohort one of a pair
   against the null distribution for cohort2}
\usage{
compareClusters(MSMthresholds, allClusterRepro, pValue, actualIGP, minThresh)
}
\arguments{
\item{MSMthresholds}{A \code{data.table} containing the thresholds between
each cluster for each comparison, as returned by `calcAllMSMthresholds`.}

\item{allClusterRepro}{A \code{data.table} containing the cluster reproduction
statistics, as calculated with `calcAllClusterRepro`.}
}
\description{
Compare the consensus clustering results for cohort one of a pair
   against the null distribution for cohort2
}
