% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-compareModels.R
\name{compareModels,ModelComparison,SurvivalModel-method}
\alias{compareModels,ModelComparison,SurvivalModel-method}
\title{Compare Two \code{SurivalModel} Objects, Returing A \code{ModelComparison} Object
With Statistics Comparing the Performance of Each Model.}
\usage{
\S4method{compareModels}{ModelComparison,SurvivalModel}(model1, model2, model2Name)
}
\arguments{
\item{model1}{An object inherting from the \code{SurvivalModel} class.}

\item{model2}{Another object inherting from the \code{SurvivalModel} class}

\item{model2Name}{A \code{character} vector with the name of the second model.}
}
\value{
A \code{ModelComparison} object with statistics comparing the two models.
}
\description{
Compare Two \code{SurivalModel} Objects, Returing A \code{ModelComparison} Object
With Statistics Comparing the Performance of Each Model.
}
\examples{
data(samplePCOSPmodel)
data(sampleClinicalModel)
data(samplePCSIsurvExp)

# Train the models
trainedPCOSPmodel <-trainModel(samplePCOSPmodel, numModels=5, minAccuracy=0.6)
trainedClinicalModel <- trainModel(sampleClinicalModel)

# Predict risk/risk-class
PCOSPpredPCSI <- predictClasses(samplePCSIsurvExp, model=trainedPCOSPmodel)
ClinicalPredPCSI <- predictClasses(samplePCSIsurvExp,
  model=trainedClinicalModel)

# Validate the models
validatedPCOSPmodel <- validateModel(trainedPCOSPmodel,
  valData=PCOSPpredPCSI)
validatedClinicalModel <- validateModel(trainedClinicalModel,
  valData=ClinicalPredPCSI)

# Compare the models
modelComp <- compareModels(validatedPCOSPmodel, validatedClinicalModel)
head(modelComp)

# Compare model comparison to another model
modelCompComp <- compareModels(modelComp, validatedPCOSPmodel)

}
