% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compassSubtypeWaterfall.R
\name{compassPlotBiomarker}
\alias{compassPlotBiomarker}
\title{Boxplot expression comparisons for the specified biomarker and subtypes}
\usage{
compassPlotBiomarker(subtypes, biomarker, exprMat)
}
\arguments{
\item{subtypes}{A \code{character} vector containing the subtype for each
sample in the expression matrix.}

\item{biomarker}{A \code{character} vector specifying the biomarker
to compare between subtypes}

\item{exprMat}{A \code{matrix} with X by Y gene expression data}

\item{palette}{A \code{character} vector specifying the name of the
RColorBrewer palette to use for the plot. Passed to `ggpubr::ggboxplot`
as the `palette` argument.}
}
\value{
A \code{ggplot} object containing a boxplot of log2 expression
   for the specified biomarker in each subtype.
}
\description{
Boxplot expression comparisons for the specified biomarker and subtypes
}
