% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compassSubtypeWaterfall.R
\name{compassPlotBiomarkers}
\alias{compassPlotBiomarkers}
\title{Boxplot the log2 expression for each biomarker between meta-clases}
\usage{
compassPlotBiomarkers(
  compassLogExprMat,
  sampClassPredwSurvivalDT,
  biomarkers,
  saveDir,
  fileName
)
}
\arguments{
\item{compassLogExprMat}{A gene by sample \code{matrix} of normalized
log2 expression values.}

\item{sampClassPredwSurvivalDT}{A \code{data.table} of cohort meta-class
predictions and survival data merged by sample.}

\item{biomarkers}{A \code{character} vector of biomarker gene names, with
each gene named for the meta-cluster it is a biomarker in.}

\item{saveDir}{An optional \code{character} vector with the path to
the directory in which the plot should be saved.}

\item{fileName}{An optional \code{character} vector specifying the name
and extension of the file. Only used if `saveDir` is also specified.
Saving is done via the `ggsave` function from `ggplot2`.}
}
\value{
A \code{gglot} object containing a plot grid with a boxplot
  for each specified biomarker.
}
\description{
Boxplot the log2 expression for each biomarker between meta-clases
}
