% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSampleClassifier.R
\name{computeCohortCommMat}
\alias{computeCohortCommMat}
\title{Subset all cohort expression data to `metaGenes`, cbind them and return a gene by
   sample expression matrix with missing values imputed using
   `impute::impute.knn`.}
\usage{
computeCohortCommMat(cohortsDataL, metaGenes)
}
\arguments{
\item{cohortsDataL}{A \code{list} per cohort gene by sample gene expression
matrixes.}

\item{metaGenes}{A \code{character} vector of meta genes to subset on, as
returned by `getMetaGenes`.}
}
\value{
A \code{matrix} of gene by sample expression data, with the rows
   containing `metaGenes` and the columns containing sample from all cohorts.
}
\description{
Subset all cohort expression data to `metaGenes`, cbind them and return a gene by
   sample expression matrix with missing values imputed using
   `impute::impute.knn`.
}
