% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerCellineDrugSensitivities.R
\name{computeConcIndOnSharedDrugs}
\alias{computeConcIndOnSharedDrugs}
\title{Compute concordance index between all datasets for a specific predicted
    subtype/meta-class.}
\usage{
computeConcIndOnSharedDrugs(mergedDT, class, na.rm = TRUE, method = "noether")
}
\arguments{
\item{mergedDT}{A \code{data.table} with merged cell-line class predictions
and cell-line drug sensitivity, as returned by the `mergeClassandDrugs`
function in this paackage.}

\item{class}{A \code{character} vecotr specifying the meta-class/subtype
to compute concordance index for.}

\item{na.rm}{A \code{boolean} specifying whether to
remove NA's from the data when calculating concordance index.}

\item{method}{A \code{character} vector specifying the method to use
for calculating concordance index. Passed to `survcomp::concordance.index`
as the `method` argument.}
}
\description{
Compute concordance index between all datasets for a specific predicted
    subtype/meta-class.
}
