% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerComparison.R
\name{computeGeneBiomarkerScores}
\alias{computeGeneBiomarkerScores}
\title{Compute the biomarker signature score for a gene}
\usage{
computeGeneBiomarkerScores(cohortsDataL, annotSampMetaClassDT, geneName)
}
\arguments{
\item{cohortsDataL}{A \code{list} per cohort gene by sample gene expression
matrixes.}

\item{annotSampMetaClassDT}{A \code{data.table} containing the per cohort per
labelled sample meta-class predictions. As returned by the
`annotateSampleMetaClasses` function in this package.}

\item{geneName}{A \code{character} vector with the name of the biomarker
gene to score.}

\item{clusterLabels}{An optional \code{character} vector to use when
annotating the sample meta-classes in the biomarker score data.table.
If excluded uses the unique meta-class names from sampleMetaClassDT,
sorted lexically.}
}
\value{
A \code{data.table} with the columns 'cohorts', 'samples',
    'metaClasses', and 'expression' containing the expression value for each
    cohort for the sspecified gene.
}
\description{
Compute the biomarker signature score for a gene
}
