% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalysis.R
\name{computePathwayScores}
\alias{computePathwayScores}
\title{Calculate the pathway score for a list of pathway specific gene-sets for
   the top 1000 genes from each meta-class.}
\usage{
computePathwayScores(rankedMetaClassGenes, pathwayL, referenceGenes)
}
\arguments{
\item{rankedMetaClassGenes}{A \code{data.table} with the top 1000 genes per
meta-class along with the associated weighted mean of the effect size and
meta-class annotations.}

\item{pathwayL}{A \code{list} of `GSC` objects from the `piano` package,
as returned by the `loadGnSetGMTs` function in this package.}

\item{referenceGenes}{A \code{character} vector with the names of all common
genes in the gene exoression cohorts.}
}
\value{
A \code{data.table} with the pathway score and p-value for each
  gene set in each pathway by mete-class.
}
\description{
Calculate the pathway score for a list of pathway specific gene-sets for
   the top 1000 genes from each meta-class.
}
