% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSignatureComparison.R
\name{computeSigScoreDT}
\alias{computeSigScoreDT}
\title{Calculate the gene signature score}
\usage{
computeSigScoreDT(cohortsDataL, sampleMetaClassDT, signatureGenes)
}
\arguments{
\item{cohortsDataL}{A \code{list} of gene by sample cohort gene expression
matrixes.}

\item{sampleMetaClassDT}{A \code{data.table} containing the per cohort per
sample meta-class predictions.}

\item{signatureGenes}{A \code{character} vector of gene names in the
expression signature.}
}
\value{
A \code{data.table} of signature scores with the columns 'samples',
   'sigScores', 'cohorts' and 'metaClasses'.
}
\description{
Calculate the gene signature score
}
