% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerCellineDrugSensitivities.R
\name{computeWilcoxOnSharedDrugs}
\alias{computeWilcoxOnSharedDrugs}
\title{Calculate the two-sided wilcox p-values for a drug shared between two
   meta-classes/subtypes for all drugs common to all included datasets.}
\usage{
computeWilcoxOnSharedDrugs(mergedDT, class1, class0)
}
\arguments{
\item{mergedDT}{A \code{data.table} with merged cell-line class predictions
and cell-line drug sensitivity, as returned by the `mergeClassandDrugs`
function in this paackage. Note: You may want to subset out datasets
which do not have many overlapping drugs between them, as only common drugs
will be included in the test.}

\item{class1}{A \code{character} vector with the class/subtype label
to compare against.}

\item{class0}{A \code{character} vector with the class/subtype label
to compare with.}
}
\value{
A \code{lost} of p-values for a two-sided wilcox test between
  cell-lines for drugs common to all included datasets.
}
\description{
Calculate the two-sided wilcox p-values for a drug shared between two
   meta-classes/subtypes for all drugs common to all included datasets.
}
