% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{conClustAllCohorts}
\alias{conClustAllCohorts}
\title{Consensus cluster a list of normalized expression matixes}
\usage{
conClustAllCohorts(
  preprocCohorts,
  maxK = 5,
  distance = "pearson",
  method = "hc",
  nthread,
  reps = 1000,
  seed = NULL
)
}
\arguments{
\item{maxK}{A \code{numeric} vector containing the integer number max
 number of clusters to use. See
`?ConsensusClusterPlus` for more information.}

\item{distance}{A \code{character} vector containing the name of
the distance metric to use in `ConsensusClusterPlus`. See
`?ConsensusClusterPlus` for more information.}

\item{method}{A \code{character} vector containing the name of
the clustering method to use in `ConsensusClusterPlus`. See
`?ConsensusClusterPlus` for more information.}

\item{nthread}{An optional argument specifying the number of threads to
parallelize over. If excluded, the default number of threads from the
`BiocParallel` package will be used.}

\item{reps}{A \code{numeric} vector containing the integer number of times
to repeat clustering. Defaults to 1000.}

\item{seed}{A \code{numeric} vector containing an integer random seed to
set for sampling in the `ConsensusClustPlus` function. Defaults to NULL,
as in no seed.}

\item{cohortL}{A \code{list} of expression data matrixes, with rows
as genes and columns as samples. As retruned by the `preprocCohorts`
function.}
}
\description{
Consensus cluster a list of normalized expression matixes
}
