% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{conClustCohort}
\alias{conClustCohort}
\title{Consensus cluster on gene expression for a cohort}
\usage{
conClustCohort(
  cohort,
  maxK,
  distance,
  method,
  reps = 1000,
  suppressPlots = TRUE,
  plotTo = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{cohort}{A \code{matrix} of expression data for a cohort, with rows
as genes and columns as samples. As in a list item from
`preprocessCohorts`.}

\item{maxK}{A \code{numeric} vector containing the integer number max
 number of clusters to use. See
`?ConsensusClusterPlus` for more information.}

\item{distance}{A \code{character} vector containing the name of
the distance metric to use in `ConsensusClusterPlus`. See
`?ConsensusClusterPlus` for more information.}

\item{method}{A \code{character} vector containing the name of
the clustering method to use in `ConsensusClusterPlus`. See
`?ConsensusClusterPlus` for more information.}

\item{reps}{A \code{numeric} vector containing the integer number of times
to repeat clustering. Defaults to 1000.}

\item{suppressPlots}{A \code{logical} vector indicating whether to suppress
the plots fomr ConsensusClusterPlus. Defaults to TRUE.}

\item{plotTo}{A \code{character} vector passed to the plot argument of
`ConsensusClusterPlus`. Defaults to NULL.}

\item{seed}{A \code{numeric} vector containing an integer random seed to
set for sampling in the `ConsensusClustPlus` function.}

\item{...}{Fallthrough arguments to `ConsensusClusterPlus`. If specficied
the function defaults are overridden.}
}
\value{
A \code{list} of consensus clustering results with items optimumK,
    clusterTable, centroidClusters and classes.
}
\description{
Consensus cluster on gene expression for a cohort
}
\section{Warning}{
 This function uses random numbers, remember to set.seed()
    before running it to ensure reproducible results!
}

