% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ConsensusClusteringModel.R
\docType{class}
\name{ConsensusClusteringModel-class}
\alias{ConsensusClusteringModel-class}
\alias{.ConsensusClusteringModel}
\title{An \code{S4Model} Containing Molecular Data to be Consesus Clustered}
\description{
An \code{S4} wrapper class providing an interface to the \code{ConsensusClusterPlus}
function from the package with the same name.
}
