% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ConsensusMetaclusteringModel.R
\docType{class}
\name{ConsensusMetaclusteringModel-class}
\alias{ConsensusMetaclusteringModel-class}
\alias{.ConsensusMetaclusteringModel}
\title{An \code{S4Model} Containing Molecular Data to be Consensus Clustered}
\description{
An \code{S4} wrapper class providing an interface to the \code{ConsensusClusterPlus}
function from the package with the same name.
}
