% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calculateValidationStats.R
\name{constructMetaEstimatesDF}
\alias{constructMetaEstimatesDF}
\title{Calculate Dindex, Cindex and cohort meta-estimates from a list of cohorts
    and associated probabilities.}
\usage{
constructMetaEstimatesDF(
  probList,
  cohortData,
  seqCohorts,
  hetero = c(TRUE, FALSE, FALSE)
)
}
\arguments{
\item{probList}{A \code{list} of probability vectors, one for each cohort}

\item{cohortData}{A \code{list} of matrices or data.frames containg per cohort
survival data. Columns OS_Stats and OS are assumed to be present!}

\item{seqCohorts}{A \code{character} vector with the names of the cohorts
which contain sequencing data. Must matcht he names attribute of  probList
and cohortData.}

\item{hetero}{A \code{logocal} vector indicating if the `hetero` parameter
should be TRUE or FALSE in the combined summary, sequencing summary and
microarray summary, respectively. Defaults to TRUE, FALSE, FALSE.}
}
\value{
A \code{list} with items dIndex, cIndex, probabilities and isSequencing where
  `dIndex`/`cIndex`` contains a data.frame with statistics for each cohort and summary
  and `probabilites`/`isSequencing` contains the probabilities used in the statistical
  calcualte and wheter or not a respective cohort contains sequening data.
}
\description{
Calculate Dindex, Cindex and cohort meta-estimates from a list of cohorts
    and associated probabilities.
}
