% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calculateCohortSubtypeStats.R
\name{constructSubtypeMetaEstimatesDF}
\alias{constructSubtypeMetaEstimatesDF}
\title{Calculate stats for combined, basal and classical subtypes for each cohort and return a list of data.frames}
\usage{
constructSubtypeMetaEstimatesDF(
  probList,
  cohortData,
  cohortClasses,
  hetero = c(TRUE, TRUE, TRUE)
)
}
\arguments{
\item{probList}{A \code{list} of per cohort survival probability predictions}

\item{cohortData}{A \code{list} of per cohort survival data}

\item{cohortClasses}{A \code{list} of per cohort subtype classifications.}

\item{hetero}{A \code{logical} vector passed to `combine.est` for the
combined, basal and classical subtypes, respectively.}
}
\description{
Calculate stats for combined, basal and classical subtypes for each cohort and return a list of data.frames
}
