% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_utilities.R, R/utilities.R
\name{convertCohortToMatrix}
\alias{convertCohortToMatrix}
\title{Take in cohort data.frame and return the cohort as a numeric matrix}
\usage{
convertCohortToMatrix(cohort)

convertCohortToMatrix(cohort)
}
\arguments{
\item{cohort}{A \code{data.frame} containing cohort data}
}
\value{
A numeric \code{matrix} containing the data from the cohort
  \code{data.frame}

A numeric \code{matrix} containing the data from the cohort
  \code{data.frame}
}
\description{
Take in cohort data.frame and return the cohort as a numeric matrix

Take in cohort data.frame and return the cohort as a numeric matrix
}
