% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedClassifiers.R
\name{correlateCentroids}
\alias{correlateCentroids}
\title{Calculate the correlation between the centroids from two classifiers}
\usage{
correlateCentroids(centroid1, centroid2)
}
\arguments{
\item{centroid1}{A \code{data.table}, \code{data.frame} or \code{matrix}
containing the centroids from a classifier}

\item{centroid2}{A \code{data.table}, \code{data.frame} or \code{matrix}
containing the centroids from a classifier}
}
\value{
A \code{data.table} with the correlation the cluster centroids
    in each clustering result.
}
\description{
Calculate the correlation between the centroids from two classifiers
}
