% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyNumAberration.R
\name{countCNAscoresByProbe}
\alias{countCNAscoresByProbe}
\title{Calculate the number of CNA probes which indicate copy number gain or loss
  feat each probe, annotated with the predicted meta-class of that sample.}
\usage{
countCNAscoresByProbe(CNAscoreDT, SNPposDT, annotSampMetaClassDT)
}
\arguments{
\item{CNAscoreDT}{A \code{data.table} of probe by sample copy number calls,
where 1 represents a gain, 0 normal and -1 a loss. Probes are included
as a column since `data.table`s cannot have rownames.}

\item{SNFposDT}{A \code{data.table} with the 'Chr' holding chromosome number,
'Positon' holding the genomic position of that SNP on the chromosome and
'ID' representing the ID of that SNP probe.}

\item{annotSampleMetaClassDT}{A \code{data.table} with annotated meta-class
predictions for each sample in each cohort.}
}
\value{
A \code{data.table} containing the sample meta-class predictions and
  number of gain, loss or normal probes in all datasets.
}
\description{
Calculate the number of CNA probes which indicate copy number gain or loss
  feat each probe, annotated with the predicted meta-class of that sample.
}
