% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CoxModel.R
\docType{class}
\name{CoxModel-class}
\alias{CoxModel-class}
\alias{.CoxModel}
\title{CoxModel-class}
\description{
Fit a cox proportional hazards model (\code{survival::coxph}) to one
or more experiments inside a \code{MultiAssayExperiment} object.
}
\keyword{internal}
