% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{CSPC_MAE}
\alias{CSPC_MAE}
\title{A \code{MultiAssayExperiment} containing cohorts of pancreatic cancer
patients, for use in package examples.}
\description{
A \code{MultiAssayExperiment} containing cohorts of pancreatic cancer
patients, for use in package examples.
}
\examples{
data(CSPC_MAE)

}
