% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_densityPlot.R
\name{densityPlotModel}
\alias{densityPlotModel}
\title{Density plot AUC distribution of a PCOSP model}
\usage{
densityPlotModel(
  formattedValCohorts,
  selectedModels,
  seqCohorts,
  title,
  vlines,
  nthread,
  filePath,
  fileName
)
}
\arguments{
\item{formattedValCohorts}{A \code{list} of formatted validation cohorts}

\item{selectedModels}{A \code{list} of kTSP classificaiton models.}

\item{seqCohorts}{A \code{character} vector of names of cohorts
containing sequencing data.}

\item{title}{A \code{character} vector with the title for the
plot}

\item{vlines}{A \code{numeric} vector of length 3 indicating
where to draw a vertical line for the combine, sequencing
and microarray density plots, respectively.}

\item{nthread}{A \code{numeric} vector indicating the integer
number of threads to parallelize over.}

\item{filePath}{A \code{character} vector with the path to
save the plot to. If absent, the plot is just returned,
not saved to disk. Passed to `ggsave` function.}

\item{fileName}{A \code{character} vector with the file name
and image file extension to save the plot under. Passed
to `ggsave` function.}
}
\value{
A \code{grob} object from \code{ggplot2}. Also saves to disk if `filePath`
    and `fileName` are specified.
}
\description{
Create a density plot of the distrubtion of AUCs between the overall,
   sequencing and microarray cohorts in a PCOSP model.
}
