% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-densityPlotModelComparison.R
\name{densityPlotModelComparison}
\alias{densityPlotModelComparison}
\title{Render A Density Plot of Model Performance for an \code{S4} Object}
\usage{
densityPlotModelComparison(object, refModel, ...)
}
\arguments{
\item{object}{An \code{S4} object representing a statistical or ML model.}

\item{refModel}{An \code{S4} object representing a statistical or ML model to
compare \code{object} against.}

\item{...}{Allow additional parameters to be defined for this generic.}
}
\value{
A \code{ggplot} object containing the plot.
}
\description{
Render A Density Plot of Model Performance for an \code{S4} Object
}
