% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{.annotateConClusters}
\alias{.annotateConClusters}
\title{Helper function for `annotatedSampleMetaClasses` to convert metacluster
    indexes to names.}
\usage{
.annotateConClusters(cohort, cluster)
}
\arguments{
\item{cohort}{A \code{data.table} of non-NA per cohort metacluster identities,
as in the `data.table` returned from `getClusterwiseSamples` if subset
to a single cohort.}

\item{cluster}{A \code{list} of consensus clustering results, as retured
by the `conClustCohort` function or an item in list returned by
`conClustAllCohorts`.}
}
\description{
Helper function for `annotatedSampleMetaClasses` to convert metacluster
    indexes to names.
}
\keyword{internal}
