% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_buildRandomGeneAssignmentModel.R
\name{.buildRGAmodels}
\alias{.buildRGAmodels}
\title{Build the random gene assignment model from <paper reference>}
\usage{
.buildRGAmodels(cohortMatrix, cohortMatrixGroups, numModels)
}
\arguments{
\item{cohortMatrix}{A}

\item{numModels}{A}

\item{cohortMatrixGorups}{A}
}
\description{
Uses non-parallelized computation of the random gene assignment model to
    replicate the results from the original PCOSP paper. This method is much
    slower when multiple CPU threads are available and should only be used to
    reproduce the results of the PCOSP paper.
}
\keyword{internal}
