% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{.calcClusterCentroid}
\alias{.calcClusterCentroid}
\title{Calculate the centroid from a}
\usage{
.calcClusterCentroid(i, cohort, clusterClasses)
}
\arguments{
\item{i}{The cluster index}

\item{cohort}{The cohort data with genes as columns, rows as samples}

\item{clusterClasses}{The}
}
\description{
Calculate the centroid from a
}
\keyword{internal}
