% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_compareClinicalModels.R
\name{.calcModelAUC}
\alias{.calcModelAUC}
\title{Calculates the AUROCs for a clinical model}
\usage{
.calcModelAUC(coh, preds)
}
\arguments{
\item{coh}{A \code{matrix} or \code{data.frame} of
expression data for a clinical cohort}

\item{preds}{A \code{vector} of predicted survival
probabilities for the given cohort.}
}
\description{
Calculates the AUROCs for a clinical model
}
