% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.ceilSqrt}
\alias{.ceilSqrt}
\title{Get the ceiling of the square root of the length of a vector or list.}
\usage{
.ceilSqrt(x)
}
\arguments{
\item{x}{A vector-like object for which `length` is defined.`}
}
\description{
Get the ceiling of the square root of the length of a vector or list.
}
\keyword{internal}
