% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_densityPlot.R
\name{.densityStats}
\alias{.densityStats}
\title{Calcualte the density plot statistics for a list of validation
   cohorts and trainedd kTSP classificaiton models}
\usage{
.densityStats(formattedValCohorts, selectedModels, seqCohorts, nthread)
}
\arguments{
\item{formattedValCohorts}{A \code{list} of formatted validation cohorts}

\item{selectedModels}{A \code{list} of kTSP classificaiton models.}

\item{seqCohorts}{A \code{character} vector of names of cohorts
containing sequencing data.}

\item{nthread}{A \code{numeric} vector indicating the integer}
}
\description{
Calcualte the density plot statistics for a list of validation
   cohorts and trainedd kTSP classificaiton models
}
\keyword{interal}
