% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calcValidationStats.R
\name{.estimateConcordanceIndex}
\alias{.estimateConcordanceIndex}
\title{Estimate the concordance index for a list of validation cohorts}
\usage{
.estimateConcordanceIndex(probList, validationCohorts)
}
\arguments{
\item{probList}{A named \code{list} of survival probabilities as calculated
with `estimatePCOSPscore`}

\item{validationCohorts}{A named \code{list} of validation cohorts}
}
\value{
A \code{list} of per cohort concordance index statistics
   calculated with the `concordance.index` function from the
   `survcomp` package.
}
\description{
Estimate the concordance index for a list of validation cohorts
}
