% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calcValidationStats.R
\name{.estimateDindex}
\alias{.estimateDindex}
\title{Estimate the D-index for a list of validation cohorts}
\usage{
.estimateDindex(probList, validationCohorts)
}
\arguments{
\item{probList}{A named \code{list} of survival probabilities as calculated
with `estimatePCOSPscore`}

\item{validationCohorts}{A named \code{list} of validation cohorts}
}
\value{
A \code{list} of per cohort D index statistics
   calculated with the `D.index` function from the
   `survcomp` package.
}
\description{
Estimate the D-index for a list of validation cohorts
}
