% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_compareClinicalModels.R
\name{.estimateModelAUCs}
\alias{.estimateModelAUCs}
\title{Calculate the AUROC for a list of expression data matrices per clinical cohort}
\usage{
.estimateModelAUCs(cFeatures, cohortProbs)
}
\arguments{
\item{cFeatures}{A \code{list} of per cohort clinical data.}

\item{cohortProbs}{A \code{list} of survival probability predictions
for each cohort in `cFeatures`.}
}
\description{
Calculate the AUROC for a list of expression data matrices per clinical cohort
}
\keyword{internal}
