% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{.extractSamples}
\alias{.extractSamples}
\title{Get the sample names in a cohort cluster}
\usage{
.extractSamples(allConClusters, cohort, cluster)
}
\arguments{
\item{allConClusters}{A \code{list} of per cohort conensus clustering results,
as returned by the `conClustAllCohorts` function in this package.}

\item{cohort}{A \code{data.table} of non-NA per cohort metacluster identities,
as in the `data.table` returned from `getCohortwiseClasses` if subset
to a single cohort.}

\item{cluster}{A \code{list} of consensus clustering results, as retured
by the `conClustCohort` function or an item in list returned by
`conClustAllCohorts`.}
}
\value{
a \code{character} vector of sample names for that cohort cluster
   combination.
}
\description{
Get the sample names in a cohort cluster
}
\keyword{internal}
