% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calculateCohortSubtypeStats.R
\name{.extractSubtypeData}
\alias{.extractSubtypeData}
\title{Subset a list of probabilities and survival data to samples in a subtype}
\usage{
.extractSubtypeData(probList, cohortData, cohortSubtypes, subtype)
}
\arguments{
\item{probList}{A \code{list} of per cohort survival probability predictions}

\item{cohortData}{A \code{list} per cohort survival data}

\item{cohortSubtypes}{A \code{list} of subtypes per clinical cohort}

\item{subtype}{A \code{character} vector indicating which subtype
to subset on.}
}
\description{
Subset a list of probabilities and survival data to samples in a subtype
}
\keyword{internal}
