% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.fillVector}
\alias{.fillVector}
\title{Extend a vector to the desired length, filling with specified value}
\usage{
.fillVector(vector, desiredLength, fillVal)
}
\arguments{
\item{vector}{An input vector to extend using `fillVal``}

\item{desiredLength}{How long the output vector should be}

\item{fillVal}{Fill with NA or some value the same type as `vector`}
}
\value{
A \code{vector} of the same type as `vector` filled with `fillVal`
    up to `desiredLength`.
}
\description{
Extend a vector to the desired length, filling with specified value
}
\keyword{internal}
