% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_buildRandomGeneAssignmentModel.R
\name{.fitRGAModel}
\alias{.fitRGAModel}
\title{Train a kTSP classifier on your data}
\usage{
.fitRGAModel(n, data, labels)
}
\description{
Train a kTSP classifier on your data
}
\keyword{internal}
