% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_aucMetaEstimates.R
\name{.metaEstimateAUCstats}
\alias{.metaEstimateAUCstats}
\title{Generate a meta estimate of AUC for from a list of cohort statistics}
\usage{
.metaEstimateAUCstats(aucStats, hetero)
}
\arguments{
\item{aucStats}{A \code{list} of AUC statistics for each cohort.}

\item{hetero}{A \code{boolean} indicating whether the cohorts are contain
both sequencing and array cohorts.}
}
\value{
A \code{list} containing the meta-estimate for AUC, AUC standard
    error and the associated p-value.
}
\description{
Generate a meta estimate of AUC for from a list of cohort statistics
}
\keyword{internal}
