% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calcValidationStats.R
\name{.metaEstimateConcordanceIndex}
\alias{.metaEstimateConcordanceIndex}
\title{Meta-estimate the concordance index for a list of validation cohorts}
\usage{
.metaEstimateConcordanceIndex(concordanceIndexList, hetero)
}
\arguments{
\item{concordanceIndexList}{A \code{list} of per cohort concordance indexes}

\item{hetero}{A \code{logical} vector passed to the
hetero argument of `combine.est` from the `survcomp`
package. See `?combine.est` for more details.}
}
\description{
Meta-estimate the concordance index for a list of validation cohorts
}
