% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calcValidationStats.R
\name{.metaEstimateDindex}
\alias{.metaEstimateDindex}
\title{Meta-estimate the Dindex for a list of validation cohorts}
\usage{
.metaEstimateDindex(DindexList, hetero)
}
\arguments{
\item{DindexList}{A \code{list} of per cohort D indexes}

\item{hetero}{A \code{logical} vector passed to the
hetero argument of `combine.est` from the `survcomp`
package. See `?combine.est` for more details.}
}
\description{
Meta-estimate the Dindex for a list of validation cohorts
}
