% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{.plotSurvivalCurve}
\alias{.plotSurvivalCurve}
\title{Draw a survival curve; for use in plotSurvivalCurves to label plots by}
\usage{
.plotSurvivalCurve(survivalCurves, title, palette, inset = 0)
}
\arguments{
\item{survivalCurves}{A \code{survfit} object as returned by the
`survival::survfit` function.}

\item{title}{An optional title for the plot.}

\item{palette}{An optional RColorBrewer palette to colour the survival curves,
defaults to 'Set1'.}
}
\description{
Draw a survival curve; for use in plotSurvivalCurves to label plots by
}
\keyword{internal}
