% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_compareClinicalModels.R
\name{.predictClinicalCohortProb}
\alias{.predictClinicalCohortProb}
\title{Calculate the clinical and PCOSP probabiltie for a clinical cohort and
    a given clinical model}
\usage{
.predictClinicalCohortProb(clinicalCohort, model)
}
\arguments{
\item{model}{A \code{glm} model trained on a training cohort}

\item{clinicalCohorts}{A \code{list} of clinical cohorts to evaluate
a predictive model against}
}
\description{
Calculate the clinical and PCOSP probabiltie for a clinical cohort and
    a given clinical model
}
\keyword{internal}
