% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_utilities.R, R/utilities.R
\name{.predictKTSP}
\alias{.predictKTSP}
\title{Make predictions using trained kTSP classification model}
\usage{
.predictKTSP(formattedValCohort, selectedModels, nthread)

.predictKTSP(formattedValCohort, selectedModels, nthread)
}
\arguments{
\item{selectedModels}{A \code{list} of kTSP classificaiton models.}

\item{nthread}{A \code{numeric} vector of the integer number
of threads to parallelize over.}

\item{formattedValCohorts}{A \code{list} of formatted validation cohorts}
}
\description{
Make predictions using trained kTSP classification model

Make predictions using trained kTSP classification model
}
