% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-validateModel.R
\name{.safe_AUC}
\alias{.safe_AUC}
\title{Calculate AUC or return NAs if it fails}
\usage{
.safe_AUC(prognosis, risk, n)
}
\description{
Calculate AUC or return NAs if it fails
}
\keyword{internal}
