% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-validateModel.R
\name{.safe_concordance.index}
\alias{.safe_concordance.index}
\title{Calculate D.index or return NAs if it fails}
\usage{
.safe_concordance.index(risk, days_survived, is_deceased, n)
}
\description{
Calculate D.index or return NAs if it fails
}
\keyword{internal}
