% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-validateModel.R
\name{.safe_D.index}
\alias{.safe_D.index}
\title{Calculate concordance.index or return NAs if it fails}
\usage{
.safe_D.index(risk, days_survived, is_deceased, n)
}
\description{
Calculate concordance.index or return NAs if it fails
}
\keyword{internal}
