% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{.scaleGenewise}
\alias{.scaleGenewise}
\title{Convert to matrix, transform, scale row-wise, transform, return}
\usage{
.scaleGenewise(cohortSubset, center, scale)
}
\arguments{
\item{cohortSubet}{A \code{data.frame} containing the cohort
data subset to the genes of interest. Rows are features,
columns are samples.}
}
\description{
Convert to matrix, transform, scale row-wise, transform, return
}
\keyword{internal}
