% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{.subsetCohorts}
\alias{.subsetCohorts}
\title{Subset all cohorts in a list of}
\usage{
.subsetCohorts(cohorts, genes = TRUE, columns = TRUE)
}
\arguments{
\item{cohorts}{A \code{list} of cohorts to subset}

\item{genes}{A \code{character} vector of genes to subset to, if
excluded all genes are returned.}

\item{columns}{A \code{character} vector of samples to subset to,
if excluded all samples are returned.}
}
\description{
Subset all cohorts in a list of
}
\keyword{internal}
