% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_densityPlot.R
\name{.summarizeKTSPs}
\alias{.summarizeKTSPs}
\title{Calculate the `combine.est` for a list of kTSP model predictions
   repeatedly}
\usage{
.summarizeKTSPs(KTSPs, nModels, nthread)
}
\arguments{
\item{KTSPs}{A \code{list} kTSP classificaiton predictions
for AUC and AUC standard error.}

\item{nModels}{A \code{numeric} vector indicating the integer
number of KTSPs}
}
\description{
Calculate the `combine.est` for a list of kTSP model predictions
   repeatedly
}
