% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_utilities.R, R/utilities.R
\name{.zipLists}
\alias{.zipLists}
\title{Merge n lists element-wise into a list of sublists n items long}
\usage{
.zipLists(..., sublistNames)

.zipLists(..., sublistNames)
}
\arguments{
\item{...}{Two or more \code{list}s to zip together element-wise. New list
is named using first list.}

\item{sublistNames}{A \code{character} vector n items long, specifying the
labels for the zipped items}
}
\description{
Take n lists and combine them element-wise into a single list with each sublist
  containing the corresponding elements of the original lists. Assigns
  sublistNames as the names of the new zipped sublists.

Take n lists and combine them element-wise into a single list with each sublist
  containing the corresponding elements of the original lists. Assigns
  sublistNames as the names of the new zipped sublists.
}
\keyword{internal}
