% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_estimatePCOSPscore.R
\name{estimatePCOSPprob}
\alias{estimatePCOSPprob}
\title{Calculate the probability}
\usage{
estimatePCOSPprob(valMat, selectedModels, nthread)
}
\arguments{
\item{valMat}{A \code{matrix} of expression values for a validation
validation cohort}

\item{selectedModels}{a \code{list} of models selected using the
`buildPCOSPmodel` function.}

\item{nthread}{A \code{numeric} vector containing the integer number of
threads to parallelize the calculation across.}
}
\value{
A \code{vector} containing the probabilities per patient
}
\description{
Calculate the probability
}
