% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{existingClassifierData}
\alias{existingClassifierData}
\title{existingClassifierData}
\description{
existingClassifierData
}
\examples{
# Loads chen, birnbaum and haiderSigScores objects
data(existingClassifierData)

Three objects:
- `chen`: The genes and coefficients for the gene signature from
Chen *et al.* (2015)
- `birnbaum`: The genes and coefficients for the gene signature from
Birnbaum *et al.* (2017)
- `haiderSigScores`: The classifier risk scores from Haider *et al.* (2014)

}
