% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_utilities.R, R/utilities.R
\name{extractAllCohorts}
\alias{extractAllCohorts}
\title{Extract all cohorts from a list of cohorts}
\usage{
extractAllCohorts(cohortList, environment)

extractAllCohorts(cohortList, environment)
}
\arguments{
\item{cohortList}{A named \code{list} containing the cohorts for an analysis}

\item{environment}{The \code{environment} into which the variables will be
assigned. Use \code{globalenv()} to assign them to the global environment
(recommended).}
}
\value{
Side effects only: assigns a variable to the parent environment with
  the name of each list element appended with "_cohort"

Side effects only: assigns a variable to the parent environment with
  the name of each list element appended with "_cohort"
}
\description{
Input a named list of cohorts to be modelled and assign each cohort to an
  environment (the global environment is recommended)

Input a named list of cohorts to be modelled and assign each cohort to an
  environment (the global environment is recommended)
}
