% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{extractMetaClusters}
\alias{extractMetaClusters}
\title{Get sample metaclusters}
\usage{
extractMetaClusters(clusters)
}
\arguments{
\item{clusters}{A \code{list} of per cohort consensus clustering results,,
as returned by the `conClustAllCohorts` function in this package.}
}
\description{
Get sample metaclusters
}
