% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSignatureComparison.R
\name{extractSampleMetaClasses}
\alias{extractSampleMetaClasses}
\title{Get the per cohort sample metaclusters and return as a data.table}
\usage{
extractSampleMetaClasses(annotatedClusters)
}
\arguments{
\item{annotatedClusters}{A \code{list} of per cohort consensus clustering
results annotated with sample metaclasses, as returned by the
`annotateSampleMetaClasses` function in this package.}
}
\value{
A \code{data.table} of merge sample metaclass data with columns
   'cohort', 'samples' and 'metaClasses'
}
\description{
Get the per cohort sample metaclusters and return as a data.table
}
