% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_readGenefuFilesToSigScores.R
\name{extractSigScoresProb}
\alias{extractSigScoresProb}
\title{Get a list of `sig.score` vectors named according to the probe they were
    calculated from.}
\usage{
extractSigScoresProb(sigScores)
}
\arguments{
\item{sigScores}{A \code{list} of signature scores caluclated with `sig.score`.
As returned from the `readGenefuFilesToSigScores` function.}
}
\value{
A \code{list} of sig.score vectors with probe names as their names.
    The returned list has the same names as the sigScores list.
}
\description{
Get a list of `sig.score` vectors named according to the probe they were
    calculated from.
}
