% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{extractSurvivalData}
\alias{extractSurvivalData}
\title{Get the survival data from a list of expression}
\usage{
extractSurvivalData(cohortsDataL)
}
\arguments{
\item{cohortsDataL}{A \code{list}}
}
\value{
A \code{data.frame} with columns cohorts, sampleNames, OS and OSstatus
  containing the per sample survival information.
}
\description{
Get the survival data from a list of expression
}
