% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-extractTrainTest.R
\name{extractTrainTest,SurvivalExperiment,character-method}
\alias{extractTrainTest,SurvivalExperiment,character-method}
\title{Extract the Training and/or Testing Data as Subsets From a SurvivalExperiment}
\usage{
\S4method{extractTrainTest}{SurvivalExperiment,character}(object, extract = "both", ...)
}
\arguments{
\item{extract}{A \code{character} vector specifying whether to extract the
data labeled 'test', 'train' or to extract both as a named list.}
}
\value{

}
\description{
Extract the Training and/or Testing Data as Subsets From a SurvivalExperiment
}
