% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-extractTrainTest.R
\name{extractTrainTest}
\alias{extractTrainTest}
\title{Extract the Training and/or Testing Data as Subsets From An S4 Object.}
\usage{
extractTrainTest(object, extract, ...)
}
\arguments{
\item{object}{An \code{S4} object to extract training and testing data from.}

\item{extract}{A \code{character} vector specifying what to extract.}
}
\description{
Extract the Training and/or Testing Data as Subsets From An S4 Object.
}
