% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{findAllCohortPairs}
\alias{findAllCohortPairs}
\title{Find all non-self pair-wise combinations of cohorts}
\usage{
findAllCohortPairs(clusterNames)
}
\arguments{
\item{clusterNames}{A \code{character} vector of cohort names}
}
\value{
A \code{data.frame} with the index of all non-self pair-wise cohort combinations. Rownames
    are the names of the two clusters being compared.
}
\description{
Find all non-self pair-wise combinations of cohorts
}
