% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{findAllPossibleClusters}
\alias{findAllPossibleClusters}
\title{Find all possible clusters in all cohorts based on the optimal K value.}
\usage{
findAllPossibleClusters(allConClusters)
}
\arguments{
\item{allConClusters}{A \code{list} of consensus clustering results, as
returned by the `conClustAllCohorts` function.}
}
\value{
A \code{character} vector containing the cluster number appended to
   the cohort name.
}
\description{
Find all possible clusters in all cohorts based on the optimal K value.
}
