% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{findCohortsNotClustered}
\alias{findCohortsNotClustered}
\title{Return which cohorts were not successfully clustered in `conClustAllCohorts`}
\usage{
findCohortsNotClustered(allPossibleClusters, metaClusters)
}
\arguments{
\item{allPossibleClusters}{A \code{character} vector of cohort clusters,
as returned by the `findAllPossibleClusters` function.}

\item{metaClusters}{A \code{communities} object from the `igraph` package,
as in the `metaClusters` item in the list returned by the
`getClusterNetworkData` function.}
}
\value{
A \code{character} vector of the cohort clusters which were not
   successfully metaclustered.
}
\description{
Return which cohorts were not successfully clustered in `conClustAllCohorts`
}
