% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-findCommonGenes.R
\name{findCommonGenes,CohortList-method}
\alias{findCommonGenes,CohortList-method}
\title{Intersect Gene Names for All \code{SurvivalExperiments} in a \code{CohortList}}
\usage{
\S4method{findCommonGenes}{CohortList}(object)
}
\arguments{
\item{object}{A \code{CohortList} of \code{SurvivalExperiment}s to find common genes
between.}
}
\value{
A \code{character} vector of genes common to all \code{SurvivalExperiment}s
in the \code{CohortList}.
}
\description{
Intersect Gene Names for All \code{SurvivalExperiments} in a \code{CohortList}
}
