% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-findCommonGenes.R
\name{findCommonGenes}
\alias{findCommonGenes}
\title{Intersect Gene Names for All \code{SurvivalExperiments} in a \code{CohortList}}
\usage{
findCommonGenes(object, ...)
}
\arguments{
\item{object}{An \code{S4} object to find common genes for.}
}
\value{
A \code{character} vector of common gene names.
}
\description{
Intersect Gene Names for All \code{SurvivalExperiments} in a \code{CohortList}
}
