% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-findCommonSamples.R
\name{findCommonSamples,CohortList-method}
\alias{findCommonSamples,CohortList-method}
\title{Find Common Samples in a CohortList Object where The Columns of Each Item
Represent Samples}
\usage{
\S4method{findCommonSamples}{CohortList}(object)
}
\arguments{
\item{object}{A \code{CohortList} for which we want to find common samples
between all \code{SurvivalExperiment} objects.}
}
\value{
A \code{character} vector of common sample names.
}
\description{
Find Common Samples in a CohortList Object where The Columns of Each Item
Represent Samples
}
\examples{
data(sampleCohortList)
commonSamples <- findCommonSamples(sampleCohortList)
head(commonSamples)

}
