% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-findCommonSamples.R
\name{findCommonSamples}
\alias{findCommonSamples}
\title{Find Common Samples in a List-like S4 Object where The Columns of Each Item
Represent Samples}
\usage{
findCommonSamples(object, ...)
}
\arguments{
\item{object}{A \code{S4} object, where the columns of each element represent
samples.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
A \code{character} vector of common sample names.
}
\description{
Find Common Samples in a List-like S4 Object where The Columns of Each Item
Represent Samples
}
\examples{
data(sampleCohortList)
commonSamples <- findCommonSamples(sampleCohortList)
head(commonSamples)

}
