% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_fitGSAtoGeneSetCollection.R
\name{fitGSAtoGeneSetCollection}
\alias{fitGSAtoGeneSetCollection}
\title{Fit GSEA model to Gene Set Collection}
\usage{
fitGSAtoGeneSetCollection(PCOSPgenes, GSC, referenceGenes, adjMethod, filePath)
}
\arguments{
\item{PCOSPgenes}{A \code{data.frame} containing the PCSOP genes for the
GSEA model}

\item{GSC}{A \code{gene set collection} of data to fit the GSEA model to}

\item{referenceGenes}{A \code{data.frame}}

\item{adjMethod}{A \code{character} vector with the multiple testing
adjustment type as specified in `runGDShyper`}

\item{filePath}{An optional \code{character} vector containing the path to
a gene set collection in .gmt format. If this is excluded the \code{GSC}
parameter will be used instead.}
}
\value{
A \code{data.frame} of significant results from the model fit
}
\description{
Fit GSEA model to Gene Set Collection
}
\examples{
data(referenceGenes)
data(PCOSPgenes)
results <- fitGSEAtoGeneSetCollection()

}
