% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{fitProportionalHazardsModel}
\alias{fitProportionalHazardsModel}
\title{Fit a proportional hazards model to survival data using the `survial::coxph`
   function.}
\usage{
fitProportionalHazardsModel(metaClusterSurvAnnot)
}
\arguments{
\item{metaClusterSurvAnnot}{A \code{data.frame} containg per sample survival data
and labelled metaclusters, as returned by the `annotateMetaClusters` function
in this package.}
}
\value{
A \code{coxph} object containing the fitted proportional hazards
   model.
}
\description{
Fit a proportional hazards model to survival data using the `survial::coxph`
   function.
}
