% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{fitSurvivalCurves}
\alias{fitSurvivalCurves}
\title{Fit Kaplan-Meyer survival curves by metacluster}
\usage{
fitSurvivalCurves(metaClusterSurvAnnot)
}
\arguments{
\item{metaClusterSurvAnnot}{A \code{data.frame} or other matrix like object
containing the metacluster labels in the 'metaClusters' column,
overall survival time in the 'OS' column and survival status in the
'OSstatus' column.}
}
\description{
Fit Kaplan-Meyer survival curves by metacluster
}
