% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-forestplot.R
\name{forestplot}
\alias{forestplot}
\title{Generate a forest plot from an `S4` object}
\usage{
forestplot(object, ...)
}
\arguments{
\item{object}{An `S4` object to create a forest plot of.}

\item{...}{Allow new parameters to this generic.}
}
\value{
None, draws a forest plot.
}
\description{
Generate a forest plot from an `S4` object
}
