% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_forestPlotMetaEstimate.R
\name{forestPlotClassifierModelComparision}
\alias{forestPlotClassifierModelComparision}
\title{Draw a forest plot using the statistics calcualted in validation metaestimate}
\usage{
forestPlotClassifierModelComparision(
  classifierStats,
  stat,
  names,
  filePath,
  fileName,
  ...
)
}
\arguments{
\item{stat}{A \code{character} vector containing the statistic to
use in the forest plot. Options are "dIndex" or "cIndex".}

\item{names}{A \code{character} vector of names for each classifier in the plot.
Assumes the names are in the same order as the items in `classifierStats`.}

\item{filePath}{A \code{character} vector containing the file path to write
the plotting results to.}

\item{fileName}{A \code{character} vector containing the desired file name
with the desired file extension. Supported extensions include .pdf, .png,
.svg and .jpeg, for more information see documentation for the `ggsave()`
function from `ggplot2`.}

\item{...}{Additional arguments passed to forestplot.}

\item{validationStatsDF}{A \code{data.frame} containing statistics for Dindex
or concordance index, as available in the \code{list} returned by
`validateMetaEstimateCalculation`.}
}
\description{
Draw a forest plot using the statistics calcualted in validation metaestimate
}
\examples{

}
