% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_forestPlotMetaEstimate.R
\name{forestPlotModelComparison}
\alias{forestPlotModelComparison}
\title{Draw a forest plot using the statistics calcualted in validation metaestimate}
\usage{
forestPlotModelComparison(
  clinicalModelStats,
  stat,
  isSummary,
  filePath,
  fileName,
  ...
)
}
\arguments{
\item{stat}{A \code{character} vector containing the statistic to
use in the forest plot. Options are "dIndex" or "cIndex".}

\item{isSummary}{A \code{logical} vector indicating which cohorts contain
summary data. Must be one boolean value per row in the data.frame.}

\item{filePath}{A \code{character} vector containing the file path to write
the plotting results to.}

\item{fileName}{A \code{character} vector containing the desired file name
with the desired file extension. Supported extensions include .pdf, .png,
.svg and .jpeg, for more information see documentation for the `ggsave()`
function from `ggplot2`.}

\item{...}{Additional arguments passed to forestplot.}

\item{validationStatsDF}{A \code{data.frame} containing statistics for Dindex
or concordance index, as available in the \code{list} returned by
`validateMetaEstimateCalculation`.}
}
\description{
Draw a forest plot using the statistics calcualted in validation metaestimate
}
\examples{

}
