% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerCellineDrugSensitivities.R
\name{formatCCLsensitivityLtoDT}
\alias{formatCCLsensitivityLtoDT}
\title{Take a list of drug by cell-line drug sensitivity data.frames and merge
   them into a single `data.table``}
\usage{
formatCCLsensitivityLtoDT(CCLsensitivityL)
}
\arguments{
\item{CCLsensitivityL}{A \code{list} of per dataset drug by cell-line
drug sensitivity `data.frame`s}
}
\value{
A \code{list} of `data.table`s annotated with drugs as columns
  and celline names in the `sample` column.
}
\description{
Take a list of drug by cell-line drug sensitivity data.frames and merge
   them into a single `data.table``
}
