% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_formatValidationCohorts.R
\name{formatValidationCohorts}
\alias{formatValidationCohorts}
\title{Format the validation cohorts}
\usage{
formatValidationCohorts(validationCohorts, environment)
}
\arguments{
\item{validationCohorts}{A named \code{list} of \code{data.frame}s containing
independent patient cohorts?}

\item{environment}{The \code{environment} into which variables will be
assigned. This parameter is missing by default, which will cause the
function to return a list instead of assigning variables to an environment.}
}
\value{
A named \code{list} containing the data matrix and group labels for
  each cohort in validationCohorts or assigns variables to the specified
  environment if one \code{envronment} parameter is included.
}
\description{
Format the validation cohorts
}
\details{
We recommend using the \code{environment()} function to set the
  environment parameter. This will assign the matrix and group objects to
  your current environment as names of the cohorts with '_mat' and '_grp'
  appended to it.
}
\examples{
data(PDACexpressionData)
cohortList <- formatValidationCohorts(PDACexpressionData)

}
