% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GeneFuModel.R
\docType{class}
\name{GeneFuModel-class}
\alias{GeneFuModel-class}
\alias{.GeneFuModel}
\title{A \code{SurvivalModel} Sub-class Designed to Hold A Survival Model Generated
Using the \code{genefu} R package.}
\description{
An S4 class with a number of predefined methods for
accessing slots relavant to a surival model. More specific model types
will inherit from this class for their accessor methods and constructor.
}
