% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{getClusterNetworkData}
\alias{getClusterNetworkData}
\title{Retrieve the network graph information based on a set of significant cluster
   edges.}
\usage{
getClusterNetworkData(clusterEdges, seed = NULL)
}
\arguments{
\item{clusterEdges}{A \code{list} of significant cluster edges}

\item{seed}{An integer seed to set for for the community search.}
}
\description{
Retrieve the network graph information based on a set of significant cluster
   edges.
}
