% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{getClusterwiseSamples}
\alias{getClusterwiseSamples}
\title{Append a list column with the vector of sample names in each cohort cluster
  to the `data.table` returned by the `getCohortwiseClasses` function.}
\usage{
getClusterwiseSamples(cohortwiseClasses, allConClusters)
}
\arguments{
\item{cohortwiseClasses}{A \code{data.table} containing the meta cluster
index for each cohort cluster, as returned by the `getCohortwiseClasses`
function in this package.}

\item{allConClusters}{A \code{list} of per cohort conensus clustering results,
as returned by the `conClustAllCohorts` function in this package.}
}
\value{
A \code{data.table} equivalent to `cohortwiseClasses`, except with
  a \code{list} column called samples, containing the samples for each
  cohort per cluster per metacluster.
}
\description{
Append a list column with the vector of sample names in each cohort cluster
  to the `data.table` returned by the `getCohortwiseClasses` function.
}
