% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{getCohortwiseClasses}
\alias{getCohortwiseClasses}
\title{Extract the per cohort cluster and meta cluster indexes from metaclustering
   results.}
\usage{
getCohortwiseClasses(metaClusters, notClustered)
}
\arguments{
\item{metaClusters}{A \code{communities} object from the `igraph` package,
as in the `metaClusters` item in the list returned by the
`getClusterNetworkData` function.}
}
\value{
A \code{data.table} with the columns cohort, cohortCluster and
   metaCluster columns, matching the clusters from each individual cohort
   clustering with the calculated metaclusters.
}
\description{
Extract the per cohort cluster and meta cluster indexes from metaclustering
   results.
}
