% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{getCommonGenes}
\alias{getCommonGenes}
\title{Get a list of common genes between all `SummarizedExperiment` in a list}
\usage{
getCommonGenes(cohortsDataL)
}
\arguments{
\item{cohortsDataL}{A \code{list} of `SummarizedExperiments`, on for each cohort}
}
\value{
A \code{charater} vector containing the genes common to all `SummarizedExperiments` in the list
}
\description{
Get a list of common genes between all `SummarizedExperiment` in a list
}
