% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneSymbols-methods.R
\name{getGeneSymbols,SummarizedExperiment-method}
\alias{getGeneSymbols,SummarizedExperiment-method}
\title{Extract gene symbols from a SummarizedExperiment object}
\usage{
\S4method{getGeneSymbols}{SummarizedExperiment}(dataset)
}
\arguments{
\item{dataset}{A \code{SummarizedExperiment} object to extract gene symbols from}
}
\value{
A \code{character} vector containing the gene symbol from the SummarizedExperiment
}
\description{
Extract gene symbols from a SummarizedExperiment object
}
