% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getGeneSymbols}
\alias{getGeneSymbols}
\title{Generic to retrieve gene symbols from an S4 object}
\usage{
getGeneSymbols(dataset, ...)
}
\arguments{
\item{dataset}{The S4 object to extract gene symbols from}

\item{...}{Allow addition of new parameters to this generic function}
}
\value{
A \code{character} vector with
}
\description{
Generic to retrieve gene symbols from an S4 object
}
