% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{getMetaGenes}
\alias{getMetaGenes}
\title{Get a vector of the unique top n genes for each cohort appended to the
    top m genes from the}
\usage{
getMetaGenes(cohortMADrankings, geneWiseWeigthedMADdf, n, pct)
}
\arguments{
\item{cohortMADrankings}{A \code{list} of \code{data.frames} containing
the genes, madValuesa nd ranking for each cohort. As returned by
the `rowMADdfs` function.}

\item{n}{A \code{numeric} vector with the integer number of top genes
to return from each cohort.}

\item{pct}{A \code{numeric} vector with the percentage of top genes
to return from the meta-cohort rankings.}

\item{geneWiseMADdf}{A \code{data.frame} containing the ordered gene
rankings across all cohorts, as returned by `calcGeneWiseWeightedMADdf`.}
}
\value{
A \code{character} vector of the unique top n ranked genes
    per cohort and top m genes in the meta-cohort ranking.
}
\description{
Get a vector of the unique top n genes for each cohort appended to the
    top m genes from the
}
