% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalModel.R
\name{getModelSeed,SurvivalModel-method}
\alias{getModelSeed,SurvivalModel-method}
\title{Method for retrieving the random seed used for training a specific survival
model object.}
\usage{
\S4method{getModelSeed}{SurvivalModel}(object)
}
\arguments{
\item{object}{A \code{SurvivalModel} object to get the seed from.}
}
\value{
An \code{integer} seed to be used when training the a \code{SurivalModel}.
}
\description{
This should be used to set the seed before model training to ensure
reproducible results.
}
\examples{
data(sampleSurvivalModel)
getModelSeed(sampleSurivalModel)

}
