% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalModel.R
\name{getModelSeed}
\alias{getModelSeed}
\title{Generic for retrieving the randomSeed parameter from a \code{SurvivalModel} object.}
\usage{
getModelSeed(object)
}
\arguments{
\item{object}{An \code{S4} object to get the seed from.}
}
\value{
An \code{integer} seed to be used when training the a \code{SurivvalModel}.
}
\description{
This should be used to set the seed before model training to ensure
reproducible results.
}
\examples{
data(sampleSurvivalModel)
getModelSeed(sampleSurivalModel)

}
