% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-getTopFeatures.R
\name{getTopFeatures}
\alias{getTopFeatures}
\title{Get the Top Predictive Features from an S4 Object}
\usage{
getTopFeatures(object, ...)
}
\arguments{
\item{object}{An \code{S4} object to get top scoring features from.}

\item{...}{Allow additional parameters to be defined for this generic.}
}
\value{
A \code{character} vector of top predictive features.
}
\description{
Get the Top Predictive Features from an S4 Object
}
\examples{
data(samplePCOSPmodel)

# Train the model
trainedPCOSPmodel <- trainModel(samplePCOSPmodel, numModels=10,
  minAccuracy=0.6)

# Get the top features
topFeatures <- getTopFeatures(trainedPCOSPmodel, numModels=5)

}
