% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{getTopGenes}
\alias{getTopGenes}
\title{Extract the top n genes from each cohort}
\usage{
getTopGenes(cohortMADrankings, n)
}
\arguments{
\item{cohortMADrankings}{A \code{list} of \code{data.frames} containing
the genes, madValuesa nd ranking for each cohort. As returned by
the `rowMADdfs` function.}

\item{n}{A \code{numeric} vector with the integer number of top genes
to return}
}
\value{
The `cohortMADrankings` \code{data.frame} for the top n genes
    in each cohort.
}
\description{
Extract the top n genes from each cohort
}
