% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDXdrugSubtyping.R
\name{ggarrangePlotL}
\alias{ggarrangePlotL}
\title{Take a long list of plots and chunk it into a list of plot chunck x chunk
   plot grids.}
\usage{
ggarrangePlotL(plotL, chunk)
}
\arguments{
\item{chunk}{A \code{numeric} vector indicating the integer number
of plots per plot grid. All returned grids have equal ncol and nrow (
i.e., are ~ square).}

\item{plots}{A \code{list} of grob or ggplot objects which can be
passed to the `ggpubr::ggarrange` function.}
}
\description{
Take a long list of plots and chunk it into a list of plot chunck x chunk
   plot grids.
}
