% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedClassifiers.R
\name{heatmapClassifCors}
\alias{heatmapClassifCors}
\title{Heatmap the correlation between classifier subtypes}
\usage{
heatmapClassifCors(assocStatsDT, dendro = TRUE, saveDir, fileName)
}
\arguments{
\item{dendro}{A \code{boolean} specifying whether or not to add a dendrogram
to the right side of the plot.}

\item{saveDir}{An optional \code{character} vector specifying the path
to the directory where the plot should be saved. If excluded, fileName
will not work.}

\item{fileName}{An optional \code{character} vector specifying the
name and extension of the file to save the plot it. This is passed to
the `ggplot2::ggsave`.}
}
\value{
A \code{ggplot} object with a heatmap showing the correlation
  between subtypes between each classifier.
}
\description{
Heatmap the correlation between classifier subtypes
}
