% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalysis.R
\name{heatmapPathwayScores}
\alias{heatmapPathwayScores}
\title{Heatmap the gene expression signatures for each pathway in each meta-class}
\usage{
heatmapPathwayScores(pathStatsDT, exclude, significance = 0.05)
}
\arguments{
\item{pathStatsDT}{A \code{data.table} with the pathway score and p-value for each
gene set in each pathway by mete-class. As returned by the
`computePathwayScores` function in this package.}

\item{exclude}{A \code{character} vector of meta-class labels to exclude
from the heatmap.}

\item{significance}{The alpha level to use as a cutoff for gene signatures
to be included in the heatmap.}
}
\value{
A \code{list} of `ggplot`s, one for each pathway when calculating
   the pathway statistics.
}
\description{
Heatmap the gene expression signatures for each pathway in each meta-class
}
