% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalysis.R
\name{loadGnSetGMTs}
\alias{loadGnSetGMTs}
\title{Read in all .gmt fiels within the specified directory as a list}
\usage{
loadGnSetGMTs(gmtFldir)
}
\arguments{
\item{gmtFldir}{GMT file directory, the path to the directory where
the .gmt files are located.}
}
\description{
Read in all .gmt fiels within the specified directory as a list
}
