% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMetaEffectSizeCalculation.R
\name{makeGenewiseCohortsDT}
\alias{makeGenewiseCohortsDT}
\title{Build a long data.table holding the gene expresion values for each gene in
   each sample in each cohort, labelled with the meta-class of each sample in
   each cohort.}
\usage{
makeGenewiseCohortsDT(preprocCohorts, annotSampMetaClassDT)
}
\arguments{
\item{preprocCohorts}{A \code{list} of preprocesed sample by gene cohort
expression data. As returned from the `preprocCohortL` function in this
package.}

\item{annotSampleMetaClassDT}{A \code{data.table} with annotated meta-class
predictions for each sample in each cohort.}
}
\value{
A long \code{data.table} of gene expression values per gene per sample
  per cohort, annotated with the predicted meta-class/subtype for each sample
  in each cohort.
}
\description{
Build a long data.table holding the gene expresion values for each gene in
   each sample in each cohort, labelled with the meta-class of each sample in
   each cohort.
}
