% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerCellineDrugSensitivities.R
\name{mergeClassAndDrugs}
\alias{mergeClassAndDrugs}
\title{Merge per dataset sensitivity data with per dataset predicted cell-line
   subtype/meta-class.}
\usage{
mergeClassAndDrugs(classPredL, drugSensL)
}
\arguments{
\item{classPredL}{A \code{list} of per cohort cell-line meta-class predictions,
such as a list of output from the `predictSampleMetaClass` function in
this package.}

\item{drugSensL}{A \code{list} of per dataset cell-line drug sensitivity
`data.table`s as returned by the `formatCCLsensitvityLtoDT` function
in this package.}
}
\value{
A \code{data.table} containing the columns 'celline', 'dataset',
   'predClass', 'pBasal', 'pClassical', 'pExocrine', 'drug' and 'AUC'.
}
\description{
Merge per dataset sensitivity data with per dataset predicted cell-line
   subtype/meta-class.
}
