% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_utilities.R, R/utilities.R
\name{mergeCommonData}
\alias{mergeCommonData}
\title{Merge common samples not censored before year 1}
\usage{
mergeCommonData(seqCohort, arrayCohort)

mergeCommonData(seqCohort, arrayCohort)
}
\arguments{
\item{seqCohort}{A \code{matrix} containing expression data for the sequence
cohort.}

\item{arrayCohort}{A \code{matrix} containing expression data for the array
cohort.}
}
\value{
A \code{data.frame} containing the merged cohorts

A \code{data.frame} containing the merged cohorts
}
\description{
Merge common samples not censored before year 1

Merge common samples not censored before year 1
}
