% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{mergeSurvOn}
\alias{mergeSurvOn}
\title{Merge two data.frames or data.tables on cohorts and some other shared column}
\usage{
mergeSurvOn(data1, data2, sharedColumn)
}
\arguments{
\item{sharedColumn}{A \code{character} vector containing the name
of a shared column to merge on, in addition to cohorts.}
}
\description{
Merge two data.frames or data.tables on cohorts and some other shared column
}
