% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calcValidationStats.R
\name{metaEstimateStats}
\alias{metaEstimateStats}
\title{Predict meta-estimates of Dindex and Cindex from an equal length list of
    each.}
\usage{
metaEstimateStats(DindexList, concordanceIndexList, hetero)
}
\arguments{
\item{DindexList}{A}

\item{concordanceIndexList}{A}

\item{hetero}{A}
}
\description{
Predict meta-estimates of Dindex and Cindex from an equal length list of
    each.
}
