% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{modelParams,S4Model-method}
\alias{modelParams,S4Model-method}
\title{Accessor for the Model Parameters of an \code{S4Model} Object}
\usage{
\S4method{modelParams}{S4Model}(object)
}
\arguments{
\item{object}{A \code{S4Model} Object}
}
\value{
A \code{List}- or \code{list}-like object containing all the parameters needed
to reproduce a specific model training run, including environmental
settings like the random seed and RNGkind.
}
\description{
Accessor for the Model Parameters of an \code{S4Model} Object
}
