% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{modelParams<-,S4Model,List_or_list_or_NULL-method}
\alias{modelParams<-,S4Model,List_or_list_or_NULL-method}
\title{Setter for the \code{modelParams} of an \code{S4Model}}
\usage{
\S4method{modelParams}{S4Model,List_or_list_or_NULL}(object) <- value
}
\arguments{
\item{object}{An \code{S4Model}}

\item{value}{A \code{List}- or \code{list}-like object containing the parameters
for the model.}
}
\value{
None, modifies the object.
}
\description{
Setter for the \code{modelParams} of an \code{S4Model}
}
\details{
This method if not intended for interactive use and will throw an warning
when used interactively.
}
\keyword{internal}
