% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{modelParams<-}
\alias{modelParams<-}
\title{Generic for Setting the Model Parameters of An \code{S4} Object}
\usage{
modelParams(object, ...) <- value
}
\arguments{
\item{object}{An \code{S4} Object}

\item{...}{Allow additional parameters to be defined for this generic.}

\item{value}{A \code{List}- or \code{list}-like object containing the parameters
for the model.}
}
\value{
None, modifies the object.
}
\description{
Generic for Setting the Model Parameters of An \code{S4} Object
}
\keyword{internal}
