% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PCOSP.R
\name{models,PCOSP-method}
\alias{models,PCOSP-method}
\title{Getter for the models slot of a `PCOSP` object}
\usage{
\S4method{models}{PCOSP}(object)
}
\arguments{
\item{object}{A `PCOSP` model object to retrieve the models slot from.}
}
\value{
A `SimpleList` of top scoring KTSPmodels
}
\description{
Getter for the models slot of a `PCOSP` object
}
