% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{models,S4Model-method}
\alias{models,S4Model-method}
\title{Getter for the models slot of an \code{S4Model} Object}
\usage{
\S4method{models}{S4Model}(object)
}
\arguments{
\item{object}{An \code{S4Model} object to retrieve models from.}
}
\value{
An \code{List}- or \code{list}-like object representing a model.
}
\description{
Getter for the models slot of an \code{S4Model} Object
}
