% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PCOSP.R
\name{models<-,PCOSP,SimpleList-method}
\alias{models<-,PCOSP,SimpleList-method}
\title{Setter for the models slot of a `PCOSP` object}
\usage{
\S4method{models}{PCOSP,SimpleList}(object) <- value
}
\arguments{
\item{object}{A `PCOSP` object to update}

\item{value}{A `SimpleList` of trained KTSP models}
}
\value{
None, updates the object.
}
\description{
Setter for the models slot of a `PCOSP` object
}
