% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{models<-,S4Model,List_or_list_or_NULL-method}
\alias{models<-,S4Model,List_or_list_or_NULL-method}
\title{Setter for the Models Slot of an \code{S4Model} Object}
\usage{
\S4method{models}{S4Model,List_or_list_or_NULL}(object) <- value
}
\arguments{
\item{object}{An \code{S4Model} object to retrieve models from.}

\item{value}{A \code{List}- or \code{list}-like object.}
}
\value{
An \code{List}- or \code{list}-like object representing a model.
}
\description{
Setter for the Models Slot of an \code{S4Model} Object
}
\keyword{internal}
