% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{models<-}
\alias{models<-}
\title{Accessor for the models slot of an \code{S4} object}
\usage{
models(object, ...) <- value
}
\arguments{
\item{object}{An \code{S4} object to retrieve models from.}

\item{...}{Allow new parameters to be defined for this generic.}

\item{value}{A \code{List}- or \code{list}-like object.}
}
\value{
None, updates the object.
}
\description{
Accessor for the models slot of an \code{S4} object
}
\keyword{internal}
