% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NCSModel.R
\docType{class}
\name{NCSModel-class}
\alias{NCSModel-class}
\alias{.NCSModel}
\title{S4Model Class for Metaclustering Using Network Community Search}
\description{
S4Model Class for Metaclustering Using Network Community Search
}
