% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NCSModel.R
\name{NetworkCommunitySearchModel}
\alias{NetworkCommunitySearchModel}
\alias{NCSModel}
\title{Constructor for a NetworkCommunitySearchModel (NCSModel)}
\usage{
NetworkCommunitySearchModel(model)
}
\arguments{
\item{model}{A validated \code{ConsensusMetaclusteringModel} object.}
}
\value{
An \code{NCSModel} object containing the relevant data from the
\code{ConsensusMetaclusteringModel}.
}
\description{
Use \code{igraph::fastgreedy.community} cross-cohort metaclusters for
a \code{ConsensusMetaclusteringModel} which has been validated with
\code{validateModel}.
}
