% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSignatureComparison.R
\name{normalizeCohortsList}
\alias{normalizeCohortsList}
\title{Z-score normalize a list of per cohort expression data}
\usage{
normalizeCohortsList(cohortsList)
}
\arguments{
\item{cohortsList}{A \code{list} of gene by sample cohort gene expression
matrixes.}
}
\description{
Z-score normalize a list of per cohort expression data
}
