% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{normalsMAE}
\alias{normalsMAE}
\title{A \code{MultiAssayExperiment} containing cohorts of normal patients, for
package examples.}
\description{
A \code{MultiAssayExperiment} containing cohorts of normal patients, for
package examples.
}
\examples{
data(normalsMAE)

}
