% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PCOSP.R
\docType{class}
\name{PCOSP-class}
\alias{PCOSP-class}
\alias{.PCOSP}
\title{Pancreatic Cancer Overall Survival Predictor (PCOSP) Class}
\description{
Pancreatic Cancer Overall Survival Predictor (PCOSP) Class
}
\section{Slots}{

\describe{
\item{\code{trainCohorts}}{A \code{CohortList} of \code{SurvivalExperiment}s for training the
PCOSP model.}

\item{\code{models}}{\code{SimpleList}}
}}

