% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PCOSP.R
\name{PCOSP}
\alias{PCOSP}
\title{Pancreatic Cancer Overall Survival Predictor (PCOSP) Constructor}
\usage{
PCOSP(trainCohorts)
}
\arguments{
\item{trainCohorts}{A \code{CohortList} of \code{SurvivalExperiment}s for training
the PCOSP model.}
}
\value{
A \code{PCOSP} object with training data in the \code{trainCohorts} slot. The
model slot will be empty until \code{trainModel} is called on the object.
}
\description{
Pancreatic Cancer Overall Survival Predictor (PCOSP) Constructor
}
