% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classUnions.R
\docType{class}
\name{PCOSP_or_RLS_or_RGA-class}
\alias{PCOSP_or_RLS_or_RGA-class}
\title{Class Union for PCOSP, RLSModel and RGAModel Types}
\description{
Class Union for PCOSP, RLSModel and RGAModel Types
}
