% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pdatk_options}
\alias{pdatk_options}
\title{Set or get options for my package}
\usage{
pdatk_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]}
pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{nthread}}{(\code{numeric};1) The number of cores to use for
     parallel computations. The default tries to read getOption('mc.cores'),
     then if that is null tried to use detectCores() - 2 and if that is
     less than 2 falls back to 1.}
}
}

