% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{pdatk_reset}
\alias{pdatk_reset}
\title{Reset PDATK Global Options to Default}
\format{
An object of class \code{list} of length 1.
}
\usage{
pdatk_reset
}
\description{
A convenient way to revert the option manager for the PDATK package back
  to it's default state.
}
\keyword{datasets}
