% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedClassifiers.R
\name{plotClassifierComparisons}
\alias{plotClassifierComparisons}
\title{Plot a heatmap showing the classification identify of each cell-line using
  each classifier.}
\usage{
plotClassifierComparisons(
  pubClassifSubtypeDT,
  allClassif = TRUE,
  saveDir,
  fileName
)
}
\arguments{
\item{pubClassifSubtypeDT}{\code{data.table} with the predicted subtype for each classifier
in each cell-line.}

\item{allClassif}{A \code{boolean} indicating whether plot only cell-line
with sybtypes in all classifiers.}

\item{saveDir}{An optional \code{character} vector specifying the path
to the directory where the plot should be saved. If excluded, fileName
will not work.}

\item{fileName}{An optional \code{character} vector specifying the
name and extension of the file to save the plot it. This is passed to
the `ggplot2::ggsave`.}
}
\value{
A \code{ggplot} object showing the classification of each cell-line
  in each classifier.
}
\description{
Plot a heatmap showing the classification identify of each cell-line using
  each classifier.
}
