% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{plotClusterNetwork}
\alias{plotClusterNetwork}
\title{Draw a plot showing the network graph of the significant cluster edges}
\usage{
plotClusterNetwork(
  clusterEdges,
  seed = NULL,
  palette = "Set1",
  clusterLabels,
  saveDir,
  fileName,
  ...
)
}
\arguments{
\item{clusterEdges}{A \code{data.table} containing the statistics for significant
inter-cohort cluster comparisons, as returned by the `compareClusters`
function in this package.}

\item{seed}{The seed to set for the graph community search.}

\item{palette}{The name of an RColorBewer palette to apply to the plot,
defaults to 'Set1'.}

\item{saveDir}{An optional \code{character} vector with the path to
the directory in which the plot should be saved.}

\item{fileName}{An optional \code{character} vector specifying the name
and extension of the file. Only used if `saveDir` is also specified.
Saving is done via the `ggsave` function from `ggplot2`.}

\item{clusterLabs}{An optional \code{character} vector of cluster names
corresponding to the numbered metaclusterings.}
}
\description{
Draw a plot showing the network graph of the significant cluster edges
}
