% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{plotCohortwiseSurvCurves}
\alias{plotCohortwiseSurvCurves}
\title{Plot a grid of per cohort survival curves}
\usage{
plotCohortwiseSurvCurves(
  metaClusterSurvAnnot,
  plot = TRUE,
  inset = 0.2,
  saveDir,
  fileName
)
}
\arguments{
\item{metaClusterSurvAnnot}{A \code{data.frame} containing the per cohort
survival data}

\item{inset}{An optional parameter specifying the how far to inset the legend
in the plot. Useful for adjusting legend position in plot grids. Defualts
to zero.}

\item{saveDir}{An optional \code{character} vector specifying the path
to the directory where the plot should be saved. If excluded, fileName
will not work.}

\item{fileName}{An optional \code{character} vector specifying the
name and extension of the file to save the plot it. This is passed to
the `ggplot2::ggsave`.}

\item{palette}{An optional RColorBrewer palette to colour the survival curves,
defaults to 'Set1'.}
}
\value{
A \code{ggplot} object
}
\description{
Plot a grid of per cohort survival curves
}
