% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotNetworkGraph.R
\name{plotNetworkGraph,NCSModel-method}
\alias{plotNetworkGraph,NCSModel-method}
\title{Plot a Network Graph for a Classified NCSModel Object}
\usage{
\S4method{plotNetworkGraph}{NCSModel}(object, ..., palette = "Set1", clusterLabels)
}
\arguments{
\item{object}{A classified \code{NCSModel} object, as returned by the
\code{predictClasses} method.}

\item{...}{Not used, force subsequent arguments to be named.}

\item{palette}{\code{character} A valid pallete for use in \code{RColourBrewer::brewer.pal}}

\item{clusterLabels}{A \code{character} vector of names for the metaclusters.
Defaults to the cluster number.}
}
\value{
A \code{ggplot} object containing the network graph, showing the
relative edge distances between each cluster in each cohort along with
the predicted metacluster label.
}
\description{
Visualize metaclusters predicted using network community search on the
consensus clustering results for a MultiAssayExperiment of patient
cohorts.
}
