% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotNetworkGraph.R
\name{plotNetworkGraph}
\alias{plotNetworkGraph}
\title{A Generic for Plotting a Network Graph From an S4 Object}
\usage{
plotNetworkGraph(object, ...)
}
\arguments{
\item{object}{An \code{S4} object with a valid plotNetworkGraph method set.}

\item{...}{Allow additional arguments to be defined for this generic.}
}
\value{
A network plot, either as an object or via side effects.
}
\description{
A Generic for Plotting a Network Graph From an S4 Object
}
