% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyNumAberration.R
\name{plotPerChromosomeCNAscores}
\alias{plotPerChromosomeCNAscores}
\title{Plot the distribution of gains/losses per genomic position for each
  chromosome.}
\usage{
plotPerChromosomeCNAscores(CNAscoreDT)
}
\arguments{
\item{CNAscoreDT}{A \code{data.table}  containing the sample meta-class
predictions and number of gain, loss or normal probes in all datasets.}
}
\value{
A \code{list} of `ggplot`s, one for each chromsome, faceted by
  predicted meta-class/subtype showing the percenage of copy gain vs
  copy losses for each probe (genomic region).
}
\description{
Plot the distribution of gains/losses per genomic position for each
  chromosome.
}
