% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotROC.R
\name{plotROC}
\alias{plotROC}
\title{Plot ROC curves for an `S4` object}
\usage{
plotROC(object, ...)
}
\arguments{
\item{object}{An `S4` object with a defined plotROC method.}

\item{...}{Allow new parameters to be added to this generic.}
}
\value{
A `ggplot` object containing the ROC curves.
}
\description{
Plot ROC curves for an `S4` object
}
