% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_plotROCcurves.R
\name{plotROCcurves}
\alias{plotROCcurves}
\title{Draw an ROC curve for each cohort in validationCohorts}
\usage{
plotROCcurves(
  formattedValidationCohorts,
  PCOSPscores,
  colours,
  legendLoc = "bottomright",
  filePath,
  fileName
)
}
\arguments{
\item{colours}{A}

\item{legendLoc}{A}

\item{filePath}{A}

\item{fileName}{A}

\item{validationCohorts}{A}

\item{validationStats}{A}
}
\value{
None, draws a plot
}
\description{
Draw an ROC curve for each cohort in validationCohorts
}
