% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSignatureComparison.R
\name{plotSigScoreL}
\alias{plotSigScoreL}
\title{Plot a grid of gene signature score boxplots}
\usage{
plotSigScoreL(sigScoreL, comparisons, palette = "Set1", saveDir, fileName)
}
\arguments{
\item{sigScoreL}{A \code{list} of signature scores as returned by the
`calcAllGeneSigScores` function in this package.}

\item{palette}{A \code{character} vector specifying the name of a palette
from RColorBrewer for the plot. Passed as `palette` argument to
`ggpubr:ggboxplot`. Defaults to 'Set1'.}

\item{saveDir}{An optional \code{character} vector specifying the path
to the directory where the plot should be saved. If excluded, fileName
will not work.}

\item{fileName}{An optional \code{character} vector specifying the
name and extension of the file to save the plot it. This is passed to
the `ggplot2::ggsave`.}

\item{comparison}{A \code{list} of character vectors containing the names
of cohorts to compare. All comparisons must be pairwise (i.e., each
character vector can have only two names).}
}
\description{
Plot a grid of gene signature score boxplots
}
