% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSignatureComparison.R
\name{plotSigScores}
\alias{plotSigScores}
\title{Draw a boxplot of the signature scores per sample per meta-class}
\usage{
plotSigScores(
  sigScoreDT,
  comparisons,
  sigScoreName,
  palette = "Set1",
  saveDir,
  fileName
)
}
\arguments{
\item{comparisons}{A \code{list} of length two character vectors specifying
the names of the metaclusters between which to compare means.}

\item{sigScoreName}{A \code{character} vector with the name of the gene
signature being plotted. Becomes the y-axis label.}

\item{palette}{A \code{character} vector specifying the name of a palette
from RColorBrewer for the plot. Passed as `palette` argument to
`ggpubr:ggboxplot`. Defaults to 'Set1'.}

\item{saveDir}{An optional \code{character} vector specifying the path
to the directory where the plot should be saved. If excluded, fileName
will not work.}

\item{fileName}{An optional \code{character} vector specifying the
name and extension of the file to save the plot it. This is passed to
the `ggplot2::ggsave`.}
}
\description{
Draw a boxplot of the signature scores per sample per meta-class
}
