% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{plotSurvivalCurves}
\alias{plotSurvivalCurves}
\title{Plot the survival curves for each predicted metacluster}
\usage{
plotSurvivalCurves(
  survivalCurves,
  title = "",
  showPlot = TRUE,
  palette = "Set1",
  inset = 0,
  saveDir,
  fileName
)
}
\arguments{
\item{survivalCurves}{A \code{list} whose first item is the `survfit` object
from curve fitting and second item is the associated pvalue for the
difference between survival curves.}

\item{title}{A \code{character} vector containing the plot title. Optional
and defaults to ''.}

\item{showPlot}{A \code{boolean} indicating wether or not to draw the plot.
For internal use in `plotCohortwiseSurvivalCurves` to prevent plotting
twice.}

\item{palette}{An optional RColorBrewer palette to colour the survival curves,
defaults to 'Set1'.}

\item{inset}{An optional parameter specifying the how far to inset the legend
in the plot. Useful for adjusting legend position in plot grids. Defualts
to zero.}

\item{saveDir}{An optional \code{character} vector specifying the path
to the directory where the plot should be saved. If excluded, fileName
will not work.}

\item{fileName}{An optional \code{character} vector specifying the
name and extension of the file to save the plot it. This is passed to
the `ggplot2::ggsave`.}
}
\description{
Plot the survival curves for each predicted metacluster
}
