% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-predictClasses.R
\name{predictClasses,CohortList,GeneFuModel-method}
\alias{predictClasses,CohortList,GeneFuModel-method}
\title{Use a Gene Signature Based Prediciton Model from the \code{genefu} Package
to Predict Signature Scores for Each Sample}
\usage{
\S4method{predictClasses}{CohortList,GeneFuModel}(object, model, ..., annot = NA)
}
\arguments{
\item{object}{A \code{CohortList} with \code{SurvivalExperiment}s to predict classes
for.}

\item{model}{A trained \code{GeneFuModel} object.}

\item{...}{Fall through parameters to \code{\link[genefu:sig.score]{genefu::sig.score}}.}

\item{annot}{The \code{annot} parameter passed to \code{\link[genefu:sig.score]{genefu::sig.score}}.
Defaults to NA, which assumes your assay rowname match the gene labels
in the model.}
}
\value{
A \code{CohortList} with the model predictions in the colData
slot as genefu_score for each \code{SurvivalExperiment}, and the
model in the metadata as predictionModel.
}
\description{
Use a Gene Signature Based Prediciton Model from the \code{genefu} Package
to Predict Signature Scores for Each Sample
}
