% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-predictClasses.R
\name{predictClasses,ConsensusClusteringModel,ANY-method}
\alias{predictClasses,ConsensusClusteringModel,ANY-method}
\title{Compute the Optimal Clustering Solution for a Trained
ConsensusClusteringModel}
\usage{
\S4method{predictClasses}{ConsensusClusteringModel,ANY}(object, model, ...)
}
\arguments{
\item{object}{A \code{MutliAssayExperiment} object}
}
\description{
Compute the optimal clustering solution out of possibilities generated
with trainModel. Assigns the cluster labels to the \code{MultiAssayExperiment}
object.
}
