% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-predictClasses.R
\name{predictClasses,NCSModel,ANY-method}
\alias{predictClasses,NCSModel,ANY-method}
\title{Predict Metacluster Labels for a NetworkCommunitySearchModel}
\usage{
\S4method{predictClasses}{NCSModel,ANY}(object)
}
\arguments{
\item{object}{A \code{NCSModel} which has been trained.}
}
\value{
The \code{object} model with
}
\description{
Predict Metacluster Labels for a NetworkCommunitySearchModel
}
