% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_densityPlot.R
\name{predictKTSPs}
\alias{predictKTSPs}
\title{Predict kTSP classification for a list of validation cohorts and
   classifier models}
\usage{
predictKTSPs(formattedValCohorts, selectedModels, nthread)
}
\arguments{
\item{formattedValCohorts}{A \code{list} of formatted validation cohorts}

\item{selectedModels}{A \code{list} of kTSP classificaiton models.}

\item{nthread}{A \code{numeric} vector with the integer number of
thread to parallelize over.}
}
\value{
A \code{list} predicted kTSP classifications
}
\description{
Predict kTSP classification for a list of validation cohorts and
   classifier models
}
