% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compassSubtypeWaterfall.R
\name{predictSampleMetaClass}
\alias{predictSampleMetaClass}
\title{Predict the metaclass and class probabilities for each sample}
\usage{
predictSampleMetaClass(
  exprTable,
  classifModel,
  topGenesDT,
  trainData,
  trainLabels
)
}
\arguments{
\item{exprTable}{A gene by sample matrix-like object containing normalized
log2 expression values for each gene in each sample.}

\item{classifModel}{A trained \code{randomForest} classifier, as returned
by the `randomForest::randomForest` function.}

\item{topGenesDT}{A \code{data.table} with the columns 'topGenes1'
and 'topGenes2' containing the top scoring pairs, 'genePairs' containing
a label for each pair, and 'classes' indicating the meta-class each gene pair
is associated with.}

\item{trainData}{A gene-pair by sample binary \code{matrix} where 1 represents
the predicted top scoring pair being correctly ordered, as returned
by the `calcTopGenes1Gt2Matrix` function in this pacakge.}

\item{trainLabels}{A \code{factor} vector with the meta-class for each
sample in the training data.}
}
\value{
A \code{data.table} with the columns 'sample', 'predClass', 'pBasal',
    'pClassical' and 'pExocrine' holding the predictions and probabilities
    for each sample's metaclass.
}
\description{
Predict the metaclass and class probabilities for each sample
}
