% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSampleClassifier.R
\name{predictSingleSampleClasses}
\alias{predictSingleSampleClasses}
\title{Predict sample metaclasses using a trained single sample classifier}
\usage{
predictSingleSampleClasses(g1Gt2Matrix, metaClassesFactor, nthread)
}
\arguments{
\item{g1Gt2Matrix}{A gene-pair by sample binary \code{matrix} where 1 represents gene1
is greater than gene2 (accurate prediction) and 0 represents gene2 less
than gene1 (an inaccurate prediction).}

\item{metaClassesFactor}{A \code{factor} vector with the meta-classes for
each sample in g1Gt2Matrix.}

\item{nthread}{A \code{numeric} vector with the integer number of threads to
parallelize over.}
}
\value{
A \code{list} of per sample metaclass predictions
}
\description{
Predict sample metaclasses using a trained single sample classifier
}
