% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{preprocCohorts}
\alias{preprocCohorts}
\title{Subset all cohorts to meta genes, transform}
\usage{
preprocCohorts(cohorts, metaGenes, center = TRUE, scale = FALSE)
}
\arguments{
\item{cohorts}{A \code{list} of cohort \code{data.frame}s or \code{matrix}es}

\item{metaGenes}{A \code{character} vector of meta genes to subset on, as
returned by `getMetaGenes`.}

\item{center}{A \code{logical} vector passed to the `center` argument of
the `base::scale`` function. Defaults to TRUE if excluded.}

\item{scale}{A \code{logical} vector passed to the `scale` argument of
the `base::scale` function. Defaults to FALSE if excluded.}
}
\description{
Subset all cohorts to meta genes, transform
}
