% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-normalize.R
\name{preprocessCaret}
\alias{preprocessCaret}
\title{Preprocess Data Using the \code{caret::preProcess} method, then return the
normalized data using \code{predict}.}
\usage{
preprocessCaret(x, ...)
}
\arguments{
\item{x}{The data to be normalized with \code{caret::preProcess} and
\code{caret::predict}.}

\item{...}{Fall through parameters to \code{caret::preProcess}. This can be used
to apply a range of different preprocessing methods from that package.}
}
\value{
\code{x} preprocessed according to the arguments in \code{...}.
}
\description{
Preprocess Data Using the \code{caret::preProcess} method, then return the
normalized data using \code{predict}.
}
\seealso{
\code{\link[caret:preProcess]{caret::preProcess}}, \code{\link[stats:predict]{stats::predict}}
}
