% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDXdrugSubtyping.R
\name{preprocPDXdata}
\alias{preprocPDXdata}
\title{Preprocess PDX gene expression data}
\usage{
preprocPDXdata(PDXdata, classifModel, topGenesDT, trainData, trainLabels)
}
\arguments{
\item{PDXdata}{A \code{data.table}}

\item{classifModel}{A trained \code{randomForest} classifier, as returned
by the `randomForest::randomForest` function.}

\item{topGenesDT}{A \code{data.table} with the columns 'topGenes1'
and 'topGenes2' containing the top scoring pairs, 'genePairs' containing
a label for each pair, and 'classes' indicating the meta-class each gene pair.
is associated with.}

\item{trainData}{A gene-pair by sample binary \code{matrix} where 1 represents
the predicted top scoring pair being correctly ordered, as returned
by the `calcTopGenes1Gt2Matrix` function in this pacakge.}

\item{trainLabels}{A \code{factor} vector with the meta-class for each
sample in the training data.}
}
\value{
A \code{data.table} of per sample per drug AAC values, labelled
   with sample name and predicted meta-class/subtype.
}
\description{
Preprocess PDX gene expression data
}
