% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurIST.R
\name{PurIST}
\alias{PurIST}
\title{PurIST}
\usage{
PurIST(dat)
}
\arguments{
\item{dat}{\code{vector} A single sample gene expression profile 
with gene symbols in names}
}
\value{
probability
}
\description{
Purity Independent Subtyping of Tumors
}
\examples{
probabibity <- PurIST(dat)

}
