% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{rankAllCohortGenesByMAD}
\alias{rankAllCohortGenesByMAD}
\title{Rank genes in a list of expression matrices by MAD}
\usage{
rankAllCohortGenesByMAD(cohortDataL, nthread)
}
\arguments{
\item{cohortDataL}{A \code{list} of cohort expression matrixes.}

\item{nthread}{A \code{numeric} vector containing the integer number of
threads to parallelize over.}
}
\value{
A \code{list} of `data.frame`s containng the gene names, MAD score
   and rankingsfor each cohort in `cohortDataL`
}
\description{
Rank genes in a list of expression matrices by MAD
}
