% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rankFeatures.R
\name{rankFeatures,MultiAssayExperiment-method}
\alias{rankFeatures,MultiAssayExperiment-method}
\title{Rank the Features in a \code{MultiAssayExperiment} Object}
\usage{
\S4method{rankFeatures}{MultiAssayExperiment}(
  object,
  FUN = "mad",
  RANK_FUN = "dense_rank",
  ...,
  descending = TRUE
)
}
\arguments{
\item{object}{A \code{MultiAssayExperiment} to rank the features in.}

\item{FUN}{A vectorized feature scoring function, such as \code{var} or \code{mad}.
Defaults to \code{mad}.}

\item{RANK_FUN}{A ranking function, such as \code{rank} or \code{dense_rank}. Defaults
to \code{dplyr::dense_rank}.}

\item{...}{Fall through arguments to \code{FUN}, such as \code{na.rm=TRUE}.}

\item{descending}{Should your rank function be called with \code{-} before the
values from \code{FUN}. Defaults to \code{TRUE}, which should be used if high values
returned from \code{FUN} are good.}
}
\value{
The \code{MultiAssayExperiment} with the item \code{featureRanks} in the object
metadata, which stores a \code{DataFrame} containing ranks accross all assays for
each unique feature and the additional columns \code{feature_score} and \code{feature_rank},
as calculated with \code{FUN} and \code{RANK_FUN}, respectively. Information
about which functions were used for each column can be found in the object
\code{mcols} in the \code{calculated_with} column.
}
\description{
Rank the Features in a \code{MultiAssayExperiment} Object
}
