% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rankFeatures.R
\name{rankFeatures}
\alias{rankFeatures}
\title{Rank the Features in a \code{S4} Object}
\usage{
rankFeatures(object, ...)
}
\arguments{
\item{object}{An \code{S4} object where rows represent features.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
The \code{object} with the features per value and ranking of the features
in the \code{rowData} slow of the object.
}
\description{
Rank the Features in a \code{S4} Object
}
\examples{
data(sampleICGCmicro)
rankFeatures(sampleICGCmicro)

}
