% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalysis.R
\name{rankMetaClassGenesByEffSize}
\alias{rankMetaClassGenesByEffSize}
\title{Get the top 1000 genes by meta-effect size within each meta-class.}
\usage{
rankMetaClassGenesByEffSize(metaEstimateStatsDT)
}
\arguments{
\item{metaEstimateStatsDT}{A gene by meta-class \code{data.table} containing the weigthed mean
of the meta-effectsize (calculated with `effsize::cohen.d`) for each
predicted subtype across all cohorts and samples. As returned from the
`calcClustMetaEstStats` function this this package.}
}
\value{
A \code{data.table} with the top 1000 genes per meta-class along
  with the associated weighted mean of the effect size and meta-class
  annotations.
}
\description{
Get the top 1000 genes by meta-effect size within each meta-class.
}
