% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_readGenefuFilesToSigScores.R
\name{readGenefuFilesToSigScores}
\alias{readGenefuFilesToSigScores}
\title{Read in .tsv files of expression data and calculate then return a list with
  a signature score for each expression file.}
\usage{
readGenefuFilesToSigScores(
  dataDir,
  fileNames,
  geneCoefFile,
  signed = FALSE,
  ...
)
}
\arguments{
\item{dataDir}{A \code{character} vector containing the path to the
data directory. We recommend using `file.path` to make your paths
platform agnostic!}

\item{fileNames}{A \code{character} vector of file names (with extensions)
to read from and extract the gene signature scores.}

\item{geneCoefFile}{A \code{character} vector with the file name of the
gene coefficients.}

\item{signed}{A \code{logical} vector indicating whether the gene coefficients
read in are signed or not. Default is FALSE}

\item{...}{Fall through parameters to `sig.scores` function from `genefu`.
Overrides the default settings.}
}
\value{
A \code{list} of signature scores objects (as returned by `sig.score`)
    per specified file, named based on the file name.
}
\description{
Read in .tsv files of expression data and calculate then return a list with
  a signature score for each expression file.
}
\section{warning}{
 this function assumes that the first column of expression
   data contains sample ID/identifiers! If it does not, please rearrange
   the columns before reading the files.
}

