% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeColDataFactorColumns}
\alias{removeColDataFactorColumns}
\title{Remave any factor columns from the \code{colData} of an \code{S4} object}
\usage{
removeColDataFactorColumns(x)
}
\arguments{
\item{x}{An \code{S4} object with a \code{colData} method defined for it.}
}
\value{
\code{x} with colData factor columsn converted to either integer or
character, as appropriate.
}
\description{
Remave any factor columns from the \code{colData} of an \code{S4} object
}
\examples{
data(sampleICGCmicro)
removeColDataFactorColumns(sampleICGCmicro)

}
