% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeFactorColumns}
\alias{removeFactorColumns}
\title{Convert factor columns in a rectangular object}
\usage{
removeFactorColumns(x)
}
\arguments{
\item{x}{A list-like rectangular object such as a \code{data.frame},
\code{data.table}, or \code{DataFrame}.}
}
\value{
\code{x} with factor columns converted to either integer or character,
as appropriate.
}
\description{
Convert factor columns in a rectangular object
}
\examples{
x <- data.frame(a=factor(LETTERS[1:5], b=factor(runif(5, 0, 1))))
removeFactorColumns(x)

}
