% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{renameColDataColumns}
\alias{renameColDataColumns}
\title{Rename the columns in the \code{colData} slot, or do nothing if they don't match}
\usage{
renameColDataColumns(x, values)
}
\arguments{
\item{x}{An \code{S4} object with a \code{colData} method.}

\item{values}{A character vector where names are the existing column
names and values are the new column names.}
}
\value{
\code{x} with updated column names, if they match any existing columns.
}
\description{
Rename the columns in the \code{colData} slot, or do nothing if they don't match
}
\examples{
data(sampleICGCmicro)
renameColDataColumns(sampleICGCmicro, c())

}
