% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RandomGeneAssignmentModel.R
\docType{class}
\name{RGAModel-class}
\alias{RGAModel-class}
\alias{.RGAModel}
\title{Constructor for a SurvivalModel Object.}
\value{
A `SurvivalModel` object.
}
\description{
Constructor for a SurvivalModel Object.
}
