% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RandomLabelShufflingModel.R
\docType{class}
\name{RLSModel-class}
\alias{RLSModel-class}
\alias{.RLSModel}
\title{A Generic Container for Storing Mathematical Models of
SurvivalExperiments}
\description{
An S4 class with a number of predefined methods for
accessing slots relavant to a surival model. More specific model types
will inherit from this class for their accessor methods and constructor.
}
