% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\docType{class}
\name{S4Model-class}
\alias{S4Model-class}
\alias{.S4Model}
\title{An \code{S4} Virtual Class For the Concept of a Statistical or ML Model}
\description{
An \code{S4} Virtual Class For the Concept of a Statistical or ML Model
}
\section{Slots}{

\describe{
\item{\code{trainData}}{An object inheriting from \code{List} or \code{list} representing
the training data for the model.}

\item{\code{modelParams}}{An object inherting from \code{List} or \code{list} representing
the parameters needed to train the model.}

\item{\code{models}}{An object inheriting from \code{List} or \code{list} representing the
trained models.}

\item{\code{validationStats}}{An object inheriting from \code{DFrame} or \code{data.frame}
and storing statistics assessing model performance.}

\item{\code{validationData}}{An object inheriting \code{List} or \code{list} representing
the data used to validate or evaluate the performance of a model.}

\item{\code{mcols}}{A \code{DataFrame} or 'data.frame' of item metadata for the \code{models}
slot.}

\item{\code{metadaata}}{A \code{List} or \code{list} of model level metadata.}
}}

\keyword{internal}
