% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleClinicalModel}
\alias{sampleClinicalModel}
\title{Sample ClinicalModel Containing the ICGC micro-array cohort from
\code{MetaGxPancreas} as training data.}
\description{
Sample ClinicalModel Containing the ICGC micro-array cohort from
\code{MetaGxPancreas} as training data.
}
\examples{
data(sampleClinicalModel)

}
\seealso{
\code{MetaGxPancreas::loadPancreasDatasets}
}
