% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleCohortClasses}
\alias{sampleCohortClasses}
\title{A list of sample subtypes for the data in sampleCohortList}
\description{
A \code{list} of \code{data.frames} containing clinical subtypes for the data
in the \code{sampleCohortList} object.
}
\examples{
data(sampleCohortClasses)

}
