% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleCohortList}
\alias{sampleCohortList}
\title{A Set of Example Patient Cohorts}
\description{
A \code{CohortList} object containing sample data for the PCOSP vignette. This
data is a subset of the the Pancreas datasets available in \code{MetaGxPancreas}.
}
\seealso{
\code{MetaGxPancreas::loadPancraseDatasets}
}
