% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleICGCmicro}
\alias{sampleICGCmicro}
\title{A Sample SurvivalExperiment Containing Data from the ICGC micro-array
cohort from \code{MetaGxPancreas}}
\description{
A Sample SurvivalExperiment Containing Data from the ICGC micro-array
cohort from \code{MetaGxPancreas}
}
\examples{
data(sampleICGCmicro)

}
\seealso{
\code{MetaGxPancreas::loadPancreasDatasets}
}
