% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{samplePCOSPmodel}
\alias{samplePCOSPmodel}
\title{A Sample PCOSP Model Containing the ICGC micro-array cohort from
\code{MetaGxPancreas} as training data.}
\description{
A Sample PCOSP Model Containing the ICGC micro-array cohort from
\code{MetaGxPancreas} as training data.
}
