% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{samplePCOSPpredList}
\alias{samplePCOSPpredList}
\title{Sample CohortList with PCOSP Risk Predictions}
\description{
Sample CohortList with PCOSP Risk Predictions
}
\examples{
data(samplePCOSPpredList)

}
\seealso{
\code{MetaGxPancreas::loadPancreasDatasets}
}
