% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{samplePCSIsurvExp}
\alias{samplePCSIsurvExp}
\title{Sample SurvivalExperiment Containing the PCSI rna-sequencing cohort
from \code{MetaGxPancreas}.}
\description{
Used as validation data for modelling examples
}
\examples{
data(samplePCSIsurvExp)

}
\seealso{
\code{MetaGxPancreas::loadPancreasDatasets}
}
