% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleRGAmodel}
\alias{sampleRGAmodel}
\title{Sample RGA Model Containing the ICGC micro-array cohort from
\code{MetaGxPancreas} as training data.}
\description{
Sample RGA Model Containing the ICGC micro-array cohort from
\code{MetaGxPancreas} as training data.
}
\examples{
data(sampleRGAmodel)

}
\seealso{
\code{MetaGxPancreas::loadPancreasDatasets}
}
