% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleRLSmodel}
\alias{sampleRLSmodel}
\title{Sample RLS Model Containing the ICGC micro-array cohort from
\code{MetaGxPancreas} as training data.}
\description{
Sample RLS Model Containing the ICGC micro-array cohort from
\code{MetaGxPancreas} as training data.
}
\examples{
data(sampleRLSmodel)

}
\seealso{
\code{MetaGxPancreas::loadPancreasDatasets}
}
