% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleTrainedPCOSPmodel}
\alias{sampleTrainedPCOSPmodel}
\title{A Sample Trained PCOSP Model Containing the ICGC micro-array cohort from
\code{MetaGxPancreas} as training data.}
\description{
A Sample Trained PCOSP Model Containing the ICGC micro-array cohort from
\code{MetaGxPancreas} as training data.
}
\examples{
data(sampleTrainedPCOSPmodel)

}
\seealso{
\code{MetaGxPancreas::loadPancreasDatasets}
}
