% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleValPCOSPmodel}
\alias{sampleValPCOSPmodel}
\title{Sample Validated PCOSP Model for Plotting Examples}
\description{
Sample Validated PCOSP Model for Plotting Examples
}
\examples{
data(sampleValPCOSPmodel)

}
\seealso{
\code{MetaGxPancreas::loadPancreasDatasets}
}
