% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{show,S4Model-method}
\alias{show,S4Model-method}
\title{Show method for Classes Inheriting from `S4Model`}
\usage{
\S4method{show}{S4Model}(object)
}
\arguments{
\item{object}{A `S4Model` derivative to show.}
}
\value{
None, prints to console.
}
\description{
Show method for Classes Inheriting from `S4Model`
}
\examples{
data(CSPC_MAE)
set.seed(1987)
metaclustModel <- ConMetaclustModel(CSPC_MAE, randomSeed=1987)
metaclustModel

}
