% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-splitTrainTest.R
\name{splitTrainTest}
\alias{splitTrainTest}
\title{Split an S4 Object into Training and Testing Data}
\usage{
splitTrainTest(object, ...)
}
\arguments{
\item{object}{`S4` An S4 object to split into training and testing groups
for use building machine learning models.}

\item{...}{Allow additional parmaters to be defined for this generic.}
}
\description{
Divide an object into training and testing data using random sammpling.
}
