% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-subset.R
\name{subset,CohortList-method}
\alias{subset,CohortList-method}
\title{Subset method for a \code{CohortList}}
\usage{
\S4method{subset}{CohortList}(x, subset = TRUE, select = TRUE)
}
\arguments{
\item{x}{A \code{CohortList} object}

\item{i}{The row query. Defaults to TRUE, i.e., select all.}

\item{j}{The column query. Defaults to TRUE, i.e., select all.}
}
\value{
A \code{CohortList} containing only the rows and columns selected in i
and j, respectively.
}
\description{
Works using endoapply of \code{[} over the list \code{SurvivalExperiment}s
}
