% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{subsetCohortExprs}
\alias{subsetCohortExprs}
\title{Get expression from list of SummarizedExperiments for a specified list of genes}
\usage{
subsetCohortExprs(cohortsDataL, genes)
}
\arguments{
\item{cohortsDataL}{A \code{list} of `SummarizedExperiment` objects, one per cohort}

\item{genes}{A \code{character} vector specifying the gene symbols of genes to return}
}
\description{
Get expression from list of SummarizedExperiments for a specified list of genes
}
