% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COINCIDE.R
\name{subsetCohorts}
\alias{subsetCohorts}
\title{Subset a list of cohort SummarizedExperiments to a specified set of genes}
\usage{
subsetCohorts(cohorts, genes)

subsetCohorts(cohorts, genes)
}
\arguments{
\item{cohorts}{A \code{list} of cohorts to subset}

\item{genes}{A \code{character} vector of genes to subset to, if
excluded all genes are returned.}
}
\value{
A \code{list} of `SummarizedExperiment` objects, subset to contain only the genes in `genes`

A \code{list} of `SummarizedExperiment`s containing only the genes in `genes`
}
\description{
Subset a list of cohort SummarizedExperiments to a specified set of genes

Subset all cohorts in a list of `SummarizedExperiments` to the genes specified in `genes`
}
