% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedClassifiers.R
\name{subtypeDataLwClassifCentroidL}
\alias{subtypeDataLwClassifCentroidL}
\title{Predict the meta class for a list of datasets using a list of classifiers}
\usage{
subtypeDataLwClassifCentroidL(
  rawDataL,
  classifCentroidL,
  predMetaClassL,
  seed = NULL
)
}
\arguments{
\item{rawDataL}{A \code{list} of `data.table`s containing the raw expression
data for a set of sample, such as for an experimental cohort.}

\item{classifCentroidL}{A \code{list} of classifier centroids to}

\item{predMetaClassL}{A \code{list} of predicted meta-classes using the COINCIDE
classifier in this pacakge. Each list item should correspond to the dataset
in `rawDataL`.}

\item{seed}{An optional \code{numeric} vector containing the random seed
to use when consensus clustering data using the supplied classifier list.
If a seed was used for the initial COINCIDE classification, the same seed
should be used here to ensure comparability of the results.}
}
\value{
A \code{data.table} containing the classification results
}
\description{
Predict the meta class for a list of datasets using a list of classifiers
}
