% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedClassifiers.R
\name{subtypeWithClassifier}
\alias{subtypeWithClassifier}
\title{Use an existing classifier centroid to classify exprData}
\usage{
subtypeWithClassifier(exprData, centroid, seed = NULL, reps = 100)
}
\arguments{
\item{exprData}{A \code{matrix} of gene by cell-line gene expression values.}

\item{centroid}{A \code{matrix} of gene by subtype cluster centroids from
an existing classifer.}

\item{seed}{A \code{numeric} vector with the integer seed to set for
consensus clustering in this function.}

\item{reps}{A \code{numeric} vector with the integer number of times to
repeat consensus clustering.}
}
\value{
A \code{data.table} with the predicted subtype for each classifier
  in each cell-line.
}
\description{
Use an existing classifier centroid to classify exprData
}
