% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_compareClinicalModels.R
\name{summarizeClinicalModels}
\alias{summarizeClinicalModels}
\title{Fit a generalized linear model for each clinical cohort in a list}
\usage{
summarizeClinicalModels(
  clinicalFeatures,
  cohorts,
 
    formula = "binary_grp ~ Age + Sex +\\n                                    T_status + N + M + Grade",
  ...
)
}
\arguments{
\item{clinicalFeatures}{A list of \code{data.frame}s containing the clinical
features for each cohort}

\item{cohorts}{A \code{character} vector of cohorts to subset from
the clinicalFeatures argument. If excluded all cohorts are used.}

\item{formula}{A \code{character} vector containing the model formula
expressed as a string. If excluded defaults to
"binary_grp ~ Age + Sex + T_status +N + M + Grade".}

\item{...}{Fallthrough parameters to `glm` function. If used, each cohort
is passed as the data argument and formula as the formula. All other
parameters must be specified in `...`.}
}
\value{
A named \code{list} containing the fitted model, model summary and anova
    for each cohort (named by `cohorts` argument or as
    `names(clincalFeatures)` if cohorts is absent).
}
\description{
Fit a generalized linear model for each clinical cohort in a list
}
