% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalExperiment.R
\docType{class}
\name{SurvivalExperiment-class}
\alias{SurvivalExperiment-class}
\alias{.SurvivalExperiment}
\title{SurvivalExperiment Class}
\description{
A SummarizedExperiment with mandatory numeric survival metadata columns
\code{survival_time} and \code{event_occurred}.
}
