% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalExperiment.R
\name{SurvivalExperiment}
\alias{SurvivalExperiment}
\title{Constructor for `SurvivalExperiment` Class}
\usage{
SurvivalExperiment(
  ...,
  days_survived = "days_survived",
  is_deceased = "is_deceased",
  sumExp
)
}
\arguments{
\item{...}{`pairlist` Fall through arguments to the `SummarizedExperiment`
constructor. These are ignored if `sumExp` is specified}

\item{days_survived}{A `character` vector indicating the column name in
`colData` which contains the `integer` number of days a patient
has survived since treatment at the time of data collection. If
`is_deceased` is 1/TRUE, then this is the number of days the patient lived.}

\item{is_deceased}{A `character` vector indicating the column name in
`colData` which contains `logical` or `integer` values where 0/FALSE means
a patient is alive and 1/TRUE means a patient is deceased.}

\item{sumExp}{An optional `SummarizedExperiment` object to coerce to a
`SurvivalExperiment`. If this parameter is included, all arguments to `...`
are ignored.}
}
\description{
Builds a `SurvivalExperiment` object, which is just a wrapper for a
  `SummarizedExperiment` with mandatory survival metadata numeric columns
  `days_survived` and `is_deceased`.
}
