% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalModel.R
\name{SurvivalModel}
\alias{SurvivalModel}
\title{Constructor for a SurvivalModel Object.}
\usage{
SurvivalModel(trainCohorts, minDaysSurvived = 365, ..., randomSeed)
}
\arguments{
\item{trainCohorts}{A 'SurvivelExperiment' containing training data for
the \code{SurvivalModel} object.}

\item{minDaysSurvived}{An \code{integer} minimum number of days survived to be
classified as a 'good' prognosis.}

\item{...}{Force subsequent paramters to be named. Not used.}

\item{randomSeed}{An \code{integer} random seed to use when sampling for the
this \code{SurvialModel} object. If excluded defaults to 1234.}
}
\value{
A \code{SurvivalModel} object.
}
\description{
Constructor for a SurvivalModel Object.
}
