% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SummarizedExperiment.R
\name{toPCOSP,SummarizedExperiment-method}
\alias{toPCOSP,SummarizedExperiment-method}
\title{Method for converting SummarizedExperiment objects from MetaGxPancreas into the PCOSP cohort format}
\usage{
\S4method{toPCOSP}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{A [SummarizedExperiment] object from the MetaGxPancreas datasets; assumes the assay of interest
is the first assay in the SE and that the colData of the SE has columns `days_to_death` and `vital_status`}
}
\value{
Sample x feature [data.frame] where values are measurements for each feature in each sample, two additional
    columns, OS and OS_Status, store survival in days and survival status (1 is deceased), respectively.
}
\description{
Method for converting SummarizedExperiment objects from MetaGxPancreas into the PCOSP cohort format
}
