% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SummarizedExperiment.R
\name{toPCOSP}
\alias{toPCOSP}
\title{Generic function for converting objects to the PCOSP cohort format}
\usage{
toPCOSP(object, ...)
}
\arguments{
\item{object}{[any] An object to reformat as a PCOSP cohort}

\item{...}{[pairlist] Allow additional parameters to be dedfined on this generic}
}
\value{
Sample x feature [data.frame] where values are measurements for each feature in each sample, two additional
    columns, OS and OS_Status, store survival in days and survival status (1 is deceased), respectively.
}
\description{
Generic function for converting objects to the PCOSP cohort format
}
