% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{trainData,S4Model-method}
\alias{trainData,S4Model-method}
\title{Accessor for the Training Data in a \code{S4Model} Object}
\usage{
\S4method{trainData}{S4Model}(object)
}
\arguments{
\item{object}{An \code{S4Model} object to retrieve training data from.}
}
\value{
The training data for an \code{S4Model} Object.
}
\description{
Accessor for the Training Data in a \code{S4Model} Object
}
