% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{trainData<-}
\alias{trainData<-}
\title{Generic for Accessing the Training Data of an \code{S4} Object}
\usage{
trainData(object, ...) <- value
}
\arguments{
\item{object}{An \code{S4} object to retrieve training data from.}

\item{...}{Allow new parameters to be defined for this generic.}

\item{value}{An object to place in the objects training data slot.}
}
\value{
None, updates the object.
}
\description{
Generic for Accessing the Training Data of an \code{S4} Object
}
