% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{trainData}
\alias{trainData}
\title{Generic for Accessing the Training Data of an `S4` Object}
\usage{
trainData(object, ...)
}
\arguments{
\item{object}{An `S4` object to retrieve training data from.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
The training data for an S4 object.
}
\description{
Generic for Accessing the Training Data of an `S4` Object
}
