% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-trainModel.R
\name{trainModel,GeneFuModel-method}
\alias{trainModel,GeneFuModel-method}
\title{Train a GeneFuModel Object}
\usage{
\S4method{trainModel}{GeneFuModel}(object)
}
\arguments{
\item{object}{A \code{GeneFuModel} object to train.}
}
\value{
An error message, since we have not finished implementing this
functionality yet.
}
\description{
Train a GeneFuModel Object
}
