% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSampleClassifier.R
\name{trainSingleSampleClassifer}
\alias{trainSingleSampleClassifer}
\title{Train a K-TSP classifier based on cohort sample meta-classes}
\usage{
trainSingleSampleClassifer(cohortComMat, sampleMetaClassDT, class, maxK)
}
\arguments{
\item{cohortComMat}{A gene by sample \code{matrix} containing all cohort
samples. As returned by the `computeCohortCommMat` function from this package.}

\item{sampleMetaClassDT}{A \code{data.table} containing the per cohort per
sample meta-class predictions.}

\item{class}{A \code{character} vector specifying the meta-class to train
the single sample classifer for.}

\item{maxK}{A \code{numeric} vector with the integer maximum K to use in
`switchBox::SWAP.KTSP.Train`.}
}
\value{
A \code{list} of KTSP classifiers, one for each meta-class.
}
\description{
Train a K-TSP classifier based on cohort sample meta-classes
}
