% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRASdep.R
\name{transExp}
\alias{transExp}
\title{Transform raw TPM to log10(TPM)}
\usage{
transExp(mat)
}
\arguments{
\item{mat}{\code{matrix} A expression profile in raw TPM values}
}
\value{
Relative TPM values
}
\description{
Transform raw TPM to log10(TPM)
}
\examples{
data <- transExp(mat)

}
