% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-validateModel.R
\name{validateModel,GeneFuModel,SurvivalExperiment-method}
\alias{validateModel,GeneFuModel,SurvivalExperiment-method}
\title{Validate a \code{GenefuModel} object with a single \code{SurvivalExperiment} object.}
\usage{
\S4method{validateModel}{GeneFuModel,SurvivalExperiment}(model, valData)
}
\arguments{
\item{model}{A \code{GenefuModel} object which has been trained using
\code{trainModel}.}

\item{valData}{A \code{SurvivalExperiment} to validate the model with.}
}
\value{
The \code{GeneModel} with the validation statistics in the
\code{validationStats} slot and the validation data in the \code{validationData} slot.
}
\description{
Validate a \code{GenefuModel} object with a single \code{SurvivalExperiment} object.
}
