% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-validateModel.R
\name{validateModel,PCOSP,CohortList-method}
\alias{validateModel,PCOSP,CohortList-method}
\title{Evaluate the Performance of a List of Trained KTSP Models from a PCOSP
Model}
\usage{
\S4method{validateModel}{PCOSP,CohortList}(model, validationData, ...)
}
\arguments{
\item{model}{A \code{PCOSP} model which has been trained using \code{trainModel}.}

\item{validationData}{A \code{CohortList} containing one or more
\code{SurvivalExperiment}s. The first assay in each \code{SurvivalExperiment} will
be classified using all top scoring KTSP models in \code{models(model)}.}

\item{...}{Fallthrough arguments to \code{BiocParallel::bplapply}, use this to
configure the parallelization settings for this function. For example
to specify BPARAM.}
}
\value{

}
\description{
Evaluate the Performance of a List of Trained KTSP Models from a PCOSP
Model
}
\seealso{
\code{\link[BiocParallel:bplapply]{BiocParallel::bplapply}}, \code{\link[switchBox:SWAP.KTSP.Classify]{switchBox::SWAP.KTSP.Classify}}
}
