% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-validateModel.R
\name{validateModel}
\alias{validateModel}
\title{Perform Validation on an \code{S4} Object Respresenting a Trained Model}
\usage{
validateModel(model, validationData, ...)
}
\arguments{
\item{model}{An \code{S4} object}
}
\description{
Perform Validation on an \code{S4} Object Respresenting a Trained Model
}
