% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{validationData<-,S4Model,List_or_list_or_NULL-method}
\alias{validationData<-,S4Model,List_or_list_or_NULL-method}
\title{Setter Method for the \code{validationData} of an \code{S4Model} Object.}
\usage{
\S4method{validationData}{S4Model,List_or_list_or_NULL}(object) <- value
}
\arguments{
\item{object}{An \code{S4Model} object.}

\item{value}{A \code{List}- or \code{list}-like object containing
the new validation data for the \code{S4Model} object.}
}
\value{
None, updates the object.
}
\description{
Setter Method for the \code{validationData} of an \code{S4Model} Object.
}
\keyword{internal}
