% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalModel.R
\name{validationData,SurvivalModel-method}
\alias{validationData,SurvivalModel-method}
\title{Accessor for the \code{validationData} slot of a \code{SurvivalModel} object.}
\usage{
\S4method{validationData}{SurvivalModel}(object)
}
\arguments{
\item{object}{A \code{SurvivalModel} object.}
}
\value{
A \code{CohortList} object containing the datasets used to compute
validation statistics for this model.
}
\description{
Accessor for the \code{validationData} slot of a \code{SurvivalModel} object.
}
\examples{
data(samplePCOSPmodel)
validationData(samplePCOSPmodel)

}
