% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{validationStats,S4Model-method}
\alias{validationStats,S4Model-method}
\title{Acessor for the \code{validationStats} slot of an \code{S4Model} Object}
\usage{
\S4method{validationStats}{S4Model}(object)
}
\arguments{
\item{object}{An \code{S4} object}

\item{...}{Allow definition of new arguments to this generic.}
}
\value{
A \code{data.frame} of validation statistics for the validation data
provided to \code{validateModel} function for a given \code{S4Model} object.
}
\description{
Acessor for the \code{validationStats} slot of an \code{S4Model} Object
}
