% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{validationStats<-,S4Model,DFrame_or_data.frame_data.table_or_NULL-method}
\alias{validationStats<-,S4Model,DFrame_or_data.frame_data.table_or_NULL-method}
\title{Setter for the \code{validationStats} slot on an \code{S4Model} object}
\usage{
\S4method{validationStats}{S4Model,DFrame_or_data.frame_data.table_or_NULL}(object) <- value
}
\arguments{
\item{object}{An \code{S4Model} object.}

\item{value}{A \code{DataFrame}- or \code{data.frame}-like object of validation
statistics.}
}
\value{
None, updates the object
}
\description{
Setter for the \code{validationStats} slot on an \code{S4Model} object
}
