% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalModel.R
\name{validationStats<-,SurvivalModel,data.frame-method}
\alias{validationStats<-,SurvivalModel,data.frame-method}
\title{Setter for the validationStats slot of a \code{SurvivalModel} object with a
\code{data.frame}}
\usage{
\S4method{validationStats}{SurvivalModel,data.frame}(object) <- value
}
\arguments{
\item{object}{A \code{SurvivalModel} model.}

\item{value}{A \code{data.frame} of validation statistics for a \code{SurvivelModel}
object.}
}
\value{
None, updated the object.
}
\description{
Setter for the validationStats slot of a \code{SurvivalModel} object with a
\code{data.frame}
}
