% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{validationStats<-}
\alias{validationStats<-}
\title{Setter for the \code{validationStats} slot on an \code{S4} object}
\usage{
validationStats(object, ...) <- value
}
\arguments{
\item{object}{An \code{S4} object.}

\item{...}{Allow definition of additional parameters to this generic.}

\item{value}{A \code{DataFrame}- or \code{data.frame}-like object of validation
statistics.}
}
\value{
None, updates the object
}
\description{
Setter for the \code{validationStats} slot on an \code{S4} object
}
