% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PCOSP.R
\name{validationStats}
\alias{validationStats}
\title{Accessor for the `validationStats` slot of an `S4` object}
\usage{
validationStats(object, ...)
}
\arguments{
\item{object}{An `S4` object}

\item{...}{Allow definition of new arguments to this generic.}
}
\value{
A `data.frame` of validation statistics for the validation cohorts
  provided to `validateModel` function for a given `PCOSP` object.
}
\description{
Accessor for the `validationStats` slot of an `S4` object
}
