% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compassSubtypeWaterfall.R
\name{waterfallPlotTumorResponse}
\alias{waterfallPlotTumorResponse}
\title{Create a waterfall plot of tumour response (measured as change in volume)}
\usage{
waterfallPlotTumorResponse(
  classSurvCompDT,
  noXaxis = FALSE,
  pVal,
  saveDir,
  fileName
)
}
\arguments{
\item{classSurvCompDT}{A \code{data.table} containing two meta-classes
to compare in the waterfall plot. To tests differences between specific
drug response, also subset to the drugs of interest.}

\item{noXaxis}{A \code{boolean} indicating wheter to hide x-axis labels and
ticks.}

\item{pVal}{An optional \code{character} vector containing the p-value for
difference in survival between the classes in the plot.}

\item{saveDir}{An optional \code{character} vector specifying the path
to the directory where the plot should be saved. If excluded, fileName
will not work.}

\item{fileName}{An optional \code{character} vector specifying the
name and extension of the file to save the plot it. This is passed to
the `ggplot2::ggsave`.}
}
\value{
A \code{ggplot} object containing the watefall plot
}
\description{
Create a waterfall plot of tumour response (measured as change in volume)
}
