% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_utilities.R, R/utilities.R
\name{whichNotCensoredYearOne}
\alias{whichNotCensoredYearOne}
\title{Exclude samples censored before year 1}
\usage{
whichNotCensoredYearOne(seqCohort)

whichNotCensoredYearOne(seqCohort)
}
\arguments{
\item{seqCohort}{A \code{matrix} or \code{data.frame} with
the cohort expression and survival data. Assumes the columns
OS and OS_Status are present.}
}
\value{
A sorted vector of indices

A sorted vector of indices
}
\description{
Exclude samples censored before year 1

Exclude samples censored before year 1
}
