% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-assignSubtypes.R
\name{assignSubtypes,CohortList,list-method}
\alias{assignSubtypes,CohortList,list-method}
\title{Assign Subtype Annotations to a SurvivalExperiment Object}
\usage{
\S4method{assignSubtypes}{CohortList,list}(
  object,
  subtypes,
  ...,
  sampleCol = "sample",
  subtypeCol = "subtype"
)
}
\arguments{
\item{object}{A \code{CohortList}}

\item{subtypes}{A \code{list} of \code{data.frame} objects, one per cohort, with
to subtypes to assign to the colData slot of cohorts with a matching name.}

\item{...}{Force subsequent arguments to be named. Not used.}

\item{sampleCol}{A \code{character} vector specifying the name of the column
containing the sample names. These must match the colnames of the
\code{SurvivalExperiment}. If the sample names are the rownames of the \code{subtypes}
\code{data.frame} then set this parameter to 'rownames'. Defaults to 'sample_name'.}

\item{subtypeCol}{A \code{character} vector specifying the name of the subtype
column in the \code{subtypes} \code{data.frame} object. Defaults to 'subtype'.}
}
\value{
The \code{CohortList} with the subtypes in the \code{subtypes} column
of the colData slot and a metadata item, \code{hasSubtypes}, set to TRUE for
each \code{SurvivalExperiment}.
}
\description{
Assign Subtype Annotations to a SurvivalExperiment Object
}
