% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_compareClinicalModels.R
\name{barplotModelComparison}
\alias{barplotModelComparison}
\title{Plot a bar plot comparing models for sequencing, microarray and overall cohorts}
\usage{
barplotModelComparison(
  modelComparisonStats,
  model = 1,
  names,
  colours,
  filePath,
  fileName,
  ...
)
}
\arguments{
\item{modelComparisonStats}{A \code{list} of model comparison statistics as
returned by `compareClinicalModels`.}

\item{model}{A \code{character} vector the name of the cohort the model
was trained on, or a \code{numeric} vector with the integer index of
the model. Defaults to 1, only need this if a model was trained on
more than one cohort.}

\item{names}{A \code{character} vector of names for each paired barplot}

\item{colours}{A \code{character} vector of colours for the paired boxplots}

\item{filePath}{A}

\item{fileName}{A}

\item{...}{Fallthrough arguments to `barplot`. Default values are used if
this is exlcuded.}
}
\value{
Nothing, draws a plot
}
\description{
Plot a bar plot comparing models for sequencing, microarray and overall cohorts
}
