% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-barPlotModelComparison.R
\name{barPlotModelComparison}
\alias{barPlotModelComparison}
\title{Make A Bar Plot Comparing Perforamnce Between Two \code{S4} Obects Representing
Mathematical Models.}
\usage{
barPlotModelComparison(model1, model2, ...)
}
\arguments{
\item{model1}{An\code{S4} object containing results of a mathematical model}

\item{model2}{An \code{S4} object containing results of a different mathematical
model, but with the same or overalapping samples.}
}
\value{
A bar plot comparing some aspect of model1 and model2
}
\description{
Make A Bar Plot Comparing Perforamnce Between Two \code{S4} Obects Representing
Mathematical Models.
}
