% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlotModelComparison.R
\name{barPlotModelComparison}
\alias{barPlotModelComparison}
\title{Compare two S4 Objects Representing Mathetmatical Models}
\usage{
barPlotModelComparison(model1, model2, ...)
}
\arguments{
\item{model1}{An\code{S4} object containing results of a mathematical model}

\item{model2}{An \code{S4} object containing results of a different mathematical
model, but with the same or overalapping samples.}
}
\value{
A bar plot comparing some aspect of model1 and model2
}
\description{
Compare two S4 Objects Representing Mathetmatical Models
}
