% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_buildPCOSPmodel.R
\name{buildPCOSPmodels}
\alias{buildPCOSPmodels}
\title{Build PCOSP Model from input data}
\usage{
buildPCOSPmodels(trainingCohorts, seqCohort, numModels, nthread, saveDir)
}
\arguments{
\item{trainingCohorts}{A named [`list`] of training cohorts for which
to fit and select PCOSP models.}

\item{seqCohort}{[`character`] The names of the cohorts containing sequencing data in trainingCohorts. All other
cohorts are assumped to contain microarray data.}

\item{numModels}{[`integer`] The number of models to fit}

\item{nthread}{[`integer`] The number of threads to parallelize across}

\item{saveDir}{[`character`] A path to a directory to save the model. If you
exclude this the function will return the model object instead.}
}
\value{
[`list`] Either returns the model object or, if \code{saveDir} is
  specified it saves to disk instead and returns the path.
}
\description{
The script is used for building Pancreatic cancer overall survival predictor
  using unique 89 samples profiled using microarray and sequencing platform.
}
\section{Warning}{
 This function uses random numbers; remember to
  \code{set.seed()} before running to ensure reproducible results
}

\examples{
# To ensure reproducible results
set.seed(1987)

# Load the data
data(trainingCohorts)

# Return the object
selectedModels <- buildPCOSPmodels(trainingCohorts, numModels=10, nthread=1)

# OR Save the object to disk
buildPCOSPmodel(traingCohort, numModels=10, saveDir=tempdir())

}
