% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ClinicalModel.R
\docType{class}
\name{ClinicalModel-class}
\alias{ClinicalModel-class}
\alias{.ClinicalModel}
\title{ClinicalModel Class Definition}
\description{
An S4 class with a number of predefined methods for
accessing slots relavant to a surival model. More specific model types
will inherit from this class for their accessor methods and constructor.
}
\examples{
data(sampleICGCmicro)
set.seed(1987)
survModel <- SurvivalModel(sampleICGCmicro, minDaysSurvived=385,
  randomSeed=1987)

}
