% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-compareModels.R
\name{compareModels,ModelComparison,SurvivalModel-method}
\alias{compareModels,ModelComparison,SurvivalModel-method}
\title{Compare Two \code{SurivalModel} Objects, Returing A \code{ModelComparison} Object
With Statistics Comparing the Performance of Each Model.}
\usage{
\S4method{compareModels}{ModelComparison,SurvivalModel}(model1, model2, model2Name)
}
\arguments{
\item{model1}{An object inherting from the \code{SurvivalModel} class.}

\item{model2}{Another object inherting from the \code{SurvivalModel} class}

\item{model2Name}{A \code{character} vector with the name of the second model.}
}
\value{
A \code{ModelComparison} object with statistics comparing the two models.
}
\description{
Compare Two \code{SurivalModel} Objects, Returing A \code{ModelComparison} Object
With Statistics Comparing the Performance of Each Model.
}
