% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOSP_calcValidationStats.R
\name{constructMetaEstimatesDF}
\alias{constructMetaEstimatesDF}
\title{Calculate Dindex, Cindex and cohort meta-estimates from a list of cohorts
and associated probabilities.}
\usage{
constructMetaEstimatesDF(
  probList,
  cohortData,
  seqCohorts,
  hetero = c(TRUE, FALSE, FALSE)
)
}
\arguments{
\item{probList}{A \code{list} of probability vectors, one for each cohort}

\item{cohortData}{A \code{list} of matrices or data.frames containg per cohort
survival data. Columns OS_Stats and OS are assumed to be present!}

\item{seqCohorts}{A \code{character} vector with the names of the cohorts
which contain sequencing data. Must matcht he names attribute of  probList
and cohortData.}

\item{hetero}{A \code{logical} vector indicating if the \code{hetero} parameter
should be \code{TRUE} or \code{FALSE} in the combined summary, sequencing summary and
microarray summary, respectively. Defaults to \code{TRUE}, \code{FALSE}, \code{FALSE}.}
}
\value{
A \code{list} with items:
\itemize{
\item \code{dIndex}, \code{cIndex}: a data.frame with statistics for each cohort
and summary
\item \code{probabilites}, \code{isSequencing}: the probabilities used in the
statistical calcualte and wheter or not a respective cohort contains
sequening data.
}
}
\description{
Calculate Dindex, Cindex and cohort meta-estimates from a list of cohorts
and associated probabilities.
}
