% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dropNotCensored.R
\name{dropNotCensored}
\alias{dropNotCensored}
\title{Remove Censored Patient Samples from An \code{S4} Object.}
\usage{
dropNotCensored(object, ...)
}
\arguments{
\item{object}{An \code{S4} object containing survival data which needs to have
patients who were not censored before some criteria.}

\item{...}{Allow new parmeters to be defined for this generic.}
}
\value{
\code{S4} The object subset to only those patients which pass the
censoring criteria.
}
\description{
Remove Censored Patient Samples from An \code{S4} Object.
}
