% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{existingClassifierData}
\alias{existingClassifierData}
\title{existingClassifierData}
\value{
Three objects:
\itemize{
\item \code{chen}: The genes and coefficients for the gene signature from
Chen \emph{et al.} (2015)
\item \code{birnbaum}: The genes and coefficients for the gene signature from
Birnbaum \emph{et al.} (2017)
\item \code{haiderSigScores}: The classifier risk scores from Haider \emph{et al.} (2014)
}
}
\description{
existingClassifierData
}
\examples{
# Loads chen, birnbaum and haiderSigScores objects
data(existingClassifierData)

}
