% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCommunitySearch.R
\name{extractSurvivalData}
\alias{extractSurvivalData}
\title{Get the survival data from a list of expression}
\usage{
extractSurvivalData(
  expressionData,
  survivalData,
  survivalDataLabel = "additonal"
)
}
\arguments{
\item{expressionData}{A \code{list} of per cohort sample by gene expression
matrixes annotated witht he columns 'OS' for overall survival time and
'OS_Status' for survival event.}

\item{survivalData}{A \code{data.frame} containing additional sample
survival data.}

\item{survivalDataLabel}{A \code{character} vector with the name to use
when labelling the additional survival data. Defaults to 'additional'
if not specified. This label will identify the additional survival data
in the returned \code{data.frame}}
}
\value{
A \code{data.frame} with columns cohorts, sampleNames, OS and OSstatus
  containing the per sample survival information.
}
\description{
Get the survival data from a list of expression
}
