% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-findCommonGenes.R
\name{findCommonGenes,CohortList-method}
\alias{findCommonGenes,CohortList-method}
\title{Intersect Gene Names for All \code{SurvivalExperiments} in a \code{CohortList}}
\usage{
\S4method{findCommonGenes}{CohortList}(object)

\S4method{findCommonGenes}{CohortList}(object)
}
\arguments{
\item{object}{A \code{MultiAssayExperiment} where rownames represent genes.}
}
\value{
A \code{character} vector of genes common to all \code{SurvivalExperiment}s
in the \code{CohortList}.

A \code{character} vector of genes common to all \code{experiments}
in the \code{MutliAssayExperiment}.
}
\description{
Intersect Gene Names for All \code{SurvivalExperiments} in a \code{CohortList}

Intersect Gene Names for All \code{experiments} in a \code{MultiAssayExperiment}
}
\examples{
data(sampleCohortList)
commonGenes <- findCommonGenes(sampleCohortList)
head(commonGenes)

}
