% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-findCommonSamples.R
\name{findCommonSamples}
\alias{findCommonSamples}
\title{Find Common Samples in a List-like S4 Object where The Columns of Each Item
Represent Samples}
\usage{
findCommonSamples(object, ...)
}
\arguments{
\item{object}{A \code{S4} object, where the columns of each element represent
samples.}
}
\description{
Find Common Samples in a List-like S4 Object where The Columns of Each Item
Represent Samples
}
