% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GeneFuModel.R
\docType{class}
\name{GeneFuModel-class}
\alias{GeneFuModel-class}
\alias{.GeneFuModel}
\title{A \code{SurvivalModel} Sub-class Designed to Hold A Survival Model Generated
Using the \code{genefu} R package.}
\description{
A \code{SurvivalModel} Sub-class Designed to Hold A Survival Model Generated
Using the \code{genefu} R package.
}
