% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ModelComparison.R
\name{ModelComparison}
\alias{ModelComparison}
\title{ModelComparison Constructor}
\usage{
ModelComparison(model1, model2, ...)
}
\arguments{
\item{model1}{An object with a \code{validationStats} method which returns a
\code{data.table}. Probably this object should inherit from \code{SurvivalModel}.}

\item{model2}{An object with a \code{validationStats} method which returns a
\code{data.table}. Probably this object should inherit from \code{SurvivalModel}.}

\item{...}{Not used.}
}
\value{
A \code{ModelComparison} object, which is a soft wrapper for \code{DataFrame}
which is used for method dispatch.
}
\description{
ModelComparison Constructor
}
