% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalModel.R
\name{models<-,SurvivalModel,SimpleList-method}
\alias{models<-,SurvivalModel,SimpleList-method}
\title{Setter for the models slot of a \code{SurvivalModel} object}
\usage{
\S4method{models}{SurvivalModel,SimpleList}(object) <- value
}
\arguments{
\item{object}{A \code{SurvivalModel} object to update}

\item{value}{A \code{SimpleList} of trained KTSP models}
}
\value{
None, updates the object.
}
\description{
Setter for the models slot of a \code{SurvivalModel} object
}
\examples{
data(samplePCOSPmodel)
models(samplePCOSPmodel) <- SimpleList(model1=NA)

}
