% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{models<-}
\alias{models<-}
\title{Accessor for the models slot of an \code{S4} object}
\usage{
models(object, ...) <- value
}
\arguments{
\item{object}{An \code{S4} object to retrieve models from.}

\item{...}{Allow new parameters to be defined for this generic.}

\item{value}{A \code{List}- or \code{list}-like object.}
}
\value{
None, updates the object.
}
\description{
Accessor for the models slot of an \code{S4} object
}
\examples{
data(CSPC_MAE)
set.seed(1987)
metaclustModel <- ConMetaclustModel(CSPC_MAE, randomSeed=1987)
models(metaclustModel) <- list(model1='some_kind_of_model')

}
\keyword{internal}
