% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PCOSP.R
\name{models}
\alias{models}
\title{Accessor for the models slot of an `S4` object}
\usage{
models(object, ...)
}
\arguments{
\item{object}{An `S4` object to retrieve models from.}

\item{...}{Allow}
}
\description{
Accessor for the models slot of an `S4` object
}
