% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-normalize.R
\name{normalize,MultiAssayExperiment-method}
\alias{normalize,MultiAssayExperiment-method}
\title{Normalize the \code{assays} in a \code{MultiAssayExperiment} Object}
\usage{
\S4method{normalize}{MultiAssayExperiment}(
  object,
  MARGIN = 2,
  FUN = "caretPreprocess",
  ...,
  whichAssays = seq_len(assays(object))
)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object with assays to normalize.}

\item{MARGIN}{An \code{integer} indicating if rows (1) or columns (2) should be normalized. Defaults to
2 for columns.}

\item{FUN}{A function to normalize your data with. The function should be
vectorized and}

\item{...}{Fall through parameters to FUN.}

\item{whichAssays}{A \code{numeric} or \code{character} vector specifying the indices
of the assays to normalize.}
}
\description{
Normalize the \code{assays} in a \code{MultiAssayExperiment} Object
}
