% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-normalize.R
\name{normalize,SummarizedExperiment-method}
\alias{normalize,SummarizedExperiment-method}
\title{Normalize the \code{assays} in a \code{SummarizedExperiment} Object}
\usage{
\S4method{normalize}{SummarizedExperiment}(
  object,
  MARGIN = 2,
  FUN = "scale",
  ...,
  whichAssays = seq_len(assays(object))
)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object with assays to normalize.}

\item{MARGIN}{An \code{integer} indicating if rows (1) or columns (2) should be normalized. Defaults to
columns. Defaults to 2.}

\item{FUN}{A function to normalize your data with. The function should be
vectorized and}

\item{...}{Fall through parameters to FUN.}

\item{whichAssays}{A \code{numeric} or \code{character} vector specifying the indices
of the assays to normalize.}
}
\value{
The \code{SummarizedExperiment} with one or more of the matrices in \code{assays} normalized.
}
\description{
Normalize the \code{assays} in a \code{SummarizedExperiment} Object
}
