% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PCOSP.R
\docType{class}
\name{PCOSP-class}
\alias{PCOSP-class}
\alias{.PCOSP}
\title{Pancreatic Cancer Overall Survival Predictor (PCOSP) Class}
\value{
A `SurvivalModel` object.
}
\description{
Pancreatic Cancer Overall Survival Predictor (PCOSP) Class
}
