% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PCOSP.R
\name{PCOSP}
\alias{PCOSP}
\title{Pancreatic Cancer Overall Survival Predictor (PCOSP) Constructor}
\usage{
PCOSP(trainCohort)
}
\arguments{
\item{trainCohorts}{A \code{CohortList} of \code{SurvivalExperiment}s for training
the PCOSP model, where each cohort has a unique molecular data type. It
is recommended to remove early deaths using \code{dropNotCensored} function
before starting to build your PCOSP model.}
}
\value{
A \code{PCOSP} object with training data in the \code{trainCohorts} slot. The
model slot will be empty until \code{trainModel} is called on the object.
}
\description{
Pancreatic Cancer Overall Survival Predictor (PCOSP) Constructor
}
\details{
This function assumes there is only 1 assay per \code{SurvivalExperiment}.
}
