% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotSurvivalCurves.R
\name{plotSurvivalCurves}
\alias{plotSurvivalCurves}
\title{Generic for Plotting Survival Curves from an \code{S4} Object}
\usage{
plotSurvivalCurves(object, ...)
}
\arguments{
\item{object}{An \code{S4} object to plot survival curves for.}

\item{...}{Allow new parameters to be defined on this generic.}
}
\value{
A plot, either via side-effects or as the return value.
}
\description{
Generic for Plotting Survival Curves from an \code{S4} Object
}
