% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-predictClasses.R
\name{predictClasses,ConsensusMetaclusteringModel,ANY-method}
\alias{predictClasses,ConsensusMetaclusteringModel,ANY-method}
\title{Compute the Optimal Clustering Solution for a Trained
ConsensusMetaclusteringModel}
\usage{
\S4method{predictClasses}{ConsensusMetaclusteringModel,ANY}(object, subinterval = c(0.1, 0.9))
}
\arguments{
\item{object}{A \code{MutliAssayExperiment} object}

\item{subinterval}{A \code{numeric} vector of two float values, the first
being the lower and second being the upper limit of the subinteral
to compare cluster ambiguity over. Default is c(0.1, 0.9), i.e. comparing
the 10th and 90th percentile of cluster consensus to calculate the
ambiguity of a given clustering solution. This is the value used to
selected the optimal K value from the potential solutions for each
assay in the training data.}
}
\value{
A \code{object} \code{ConsensusMetaclusteringModel}, with class predictions
assigned to the colData of \code{trianData}
}
\description{
Compute the optimal clustering solution out of possibilities generated
with trainModel. Assigns the cluster labels to the \code{MultiAssayExperiment}
object.
}
