% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RandomGeneAssignmentModel.R
\name{RandomGeneAssignmentModel}
\alias{RandomGeneAssignmentModel}
\alias{RGAModel}
\title{A Generic Container for Storing Mathematical Models of
SurvivalExperiments}
\usage{
RandomGeneAssignmentModel(trainCohorts, minDaysSurvived = 365, ..., randomSeed)
}
\arguments{
\item{trainCohorts}{A 'SurvivelExperiment' containing training data for
the `SurvivalModel` object.}

\item{minDaysSurvived}{An `integer` minimum number of days survived to be
classified as a 'good' prognosis.}

\item{...}{Force subsequent paramters to be named. Not used.}

\item{randomSeed}{An `integer` random seed to use when sampling for the
this `SurvialModel` object. If excluded defaults to 1234.}
}
\value{
A `SurvivalModel` object.
}
\description{
An S4 class with a number of predefined methods for
accessing slots relavant to a surival model. More specific model types
will inherit from this class for their accessor methods and constructor.
}
