% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RandomLabelShufflingModel.R
\name{RandomLabelShufflingModel}
\alias{RandomLabelShufflingModel}
\alias{RLSModel}
\title{RandomLabelShufflingModel Constructor}
\usage{
RandomLabelShufflingModel(trainCohorts, minDaysSurvived = 365, ..., randomSeed)
}
\arguments{
\item{trainCohorts}{A 'SurvivelExperiment' containing training data for
the `SurvivalModel` object.}

\item{minDaysSurvived}{An `integer` minimum number of days survived to be
classified as a 'good' prognosis.}

\item{...}{Force subsequent paramters to be named. Not used.}

\item{randomSeed}{An `integer` random seed to use when sampling for the
this `SurvialModel` object. If excluded defaults to 1234.}
}
\value{
A `SurvivalModel` object.
}
\description{
RandomLabelShufflingModel Constructor
}
