% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-rankFeatures.R
\name{rankFeatures,SummarizedExperiment-method}
\alias{rankFeatures,SummarizedExperiment-method}
\title{Rank the Features in a `SummarizedExperiment` Object}
\usage{
\S4method{rankFeatures}{SummarizedExperiment}(
  object,
  FUN = "rowMads",
  RANK_FUN = "dense_rank",
  ...,
  descending = TRUE,
  assay = 1
)
}
\arguments{
\item{object}{A `SummarizedExperiment` to rank the features in.}

\item{FUN}{A row-wise summary function, such as `rowVars`, `rowMads`, etc.
defaults to `MatrixGenerics::rowMads`.}

\item{RANK_FUN}{A ranking function, such as `rank` or `dense_rank`. Defaults
to `dplyr::dense_rank`.}

\item{...}{Fall through arguments to `FUN`, such as `na.rm=TRUE`.}

\item{descending}{Should your rank function be called with `-` before the
values from `FUN`. Defaults to `TRUE`.}

\item{assay}{`integer` assay to use for the ranking, as passed to the
`SummarizedExperiment::assay` function. Defaults to the first assay.}
}
\value{
The `SummarizedExperiment` with the column `feature_score` and
  `feature_rank` in the `rowData` slot. Information about which functions
  where used for each column can be found in the object `mcols` in the
  `calculated_with` column.
}
\description{
Rank the Features in a `SummarizedExperiment` Object
}
\examples{
data(sampleICGCmicro)
rankFeatures(sampleICGCmicro, FUN='rowMads', RANK_FUN='dense_rank')

}
