% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RandomLabelShufflingModel.R
\docType{class}
\name{RLSModel-class}
\alias{RLSModel-class}
\alias{.RLSModel}
\title{Constructor for a SurvivalModel Object.}
\value{
A `SurvivalModel` object.
}
\description{
Constructor for a SurvivalModel Object.
}
