% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-trainModel.R
\name{trainModel,PCOSP-method}
\alias{trainModel,PCOSP-method}
\title{Train a PCOSP Model Based on The Data the assay \code{trainMatrix}.}
\usage{
\S4method{trainModel}{PCOSP}(object, ...)
}
\arguments{
\item{object}{A \code{PCOSP} object to train.}

\item{...}{Fall through arguments to \code{BiocParallel::bplapply}}
}
\value{
A \code{PCOSP} object with the trained model in the \code{model} slot.
}
\description{
Train a PCOSP Model Based on The Data the assay \code{trainMatrix}.
}
\details{
This function is parallelized using the nthread option from
the \code{pdatk_options()} function. Please use \code{pdatk_options(nthread=n)} to
change from the default value, which is inferred when the package
is loaded. Alternatively, you can pass in your own parallel back-end
using the BPARAM parameter, which will fall through to the bplapply
call inside this function.
}
