% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-validateModel.R
\name{validateModel,ClinicalModel,SurvivalExperiment-method}
\alias{validateModel,ClinicalModel,SurvivalExperiment-method}
\title{Validate a PCOSP model with a single SurvivalExperiment object.}
\usage{
\S4method{validateModel}{ClinicalModel,SurvivalExperiment}(model, valData)
}
\arguments{
\item{model}{A \code{ClinicalModel} object which has been trained using
\code{trainModel}.}

\item{valData}{A \code{SurvivalExperiment} to validate the model with.}
}
\value{
The \code{ClinicalModel} with the validation statistics in the
\code{validationStats} slot and the validation data in the
\code{validationData} slot.
}
\description{
Validate a PCOSP model with a single SurvivalExperiment object.
}
