% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{validationData<-}
\alias{validationData<-}
\title{Generic for setting the \code{validationData} slot on an \code{S4} object}
\usage{
validationData(object, ...) <- value
}
\arguments{
\item{object}{An \code{S4} object.}

\item{...}{Allow definition of additional parameters to this generic.}

\item{value}{A \code{data.frame} of validation statistics.}
}
\value{
None, updates the object
}
\description{
Generic for setting the \code{validationData} slot on an \code{S4} object
}
\examples{
data(CSPC_MAE)
set.seed(1987)
metaclustModel <- ConMetaclustModel(CSPC_MAE, randomSeed=1987)
validationData(metaclustModel) <- list(cohort1='This should be cohort data')

}
