% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PCOSP.R
\name{validationData}
\alias{validationData}
\title{Accessor for the `validationData` slot of an `S4` object}
\usage{
validationData(object, ...)
}
\arguments{
\item{object}{An `S4` object}

\item{...}{Allow definition of new arguments to this generic.}
}
\value{
A `CohortList` with the validation data used for the `PCOSP` model,
  or nothing if the model has not be validated.
}
\description{
Accessor for the `validationData` slot of an `S4` object
}
