% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalModel.R
\name{validationData}
\alias{validationData}
\title{Accessor for the \code{validationData} slot of an \code{S4} object}
\usage{
validationData(object, ...)
}
\arguments{
\item{object}{An \code{S4} object}

\item{...}{Allow definition of new arguments to this generic.}
}
\value{
A \code{CohortList} with the validation data used for the \code{SurvivalModel} model,
or nothing if the model has not be validated.
}
\description{
Accessor for the \code{validationData} slot of an \code{S4} object
}
\examples{
data(samplePCOSPmodel)
validationData(samplePCOSPmodel)

}
