% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-S4Model.R
\name{validationData}
\alias{validationData}
\title{Accessor for the \code{validationData} slot of an \code{S4} object}
\usage{
validationData(object, ...)
}
\arguments{
\item{object}{An \code{S4} object}

\item{...}{Allow definition of new arguments to this generic.}
}
\value{
A \code{List}- or \code{list}-like object containing one or more
sets of validation data.
}
\description{
Accessor for the \code{validationData} slot of an \code{S4} object
}
\examples{
data(CSPC_MAE)
set.seed(1987)
metaclustModel <- ConMetaclustModel(CSPC_MAE, randomSeed=1987)
validationData(metaclustModel)

}
