% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalModel.R
\name{validationStats<-}
\alias{validationStats<-}
\title{Generic for setting the `validationStats` slot on an `S4` object}
\usage{
validationStats(object, ...) <- value
}
\arguments{
\item{object}{An `S4` object.}

\item{...}{Allow definition of additional parameters to this generic.}

\item{value}{A `data.frame` of validation statistics.}
}
\value{
None, updates the object
}
\description{
Generic for setting the `validationStats` slot on an `S4` object
}
