% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SurvivalModel.R
\name{validationStats<-}
\alias{validationStats<-}
\title{Setter for the \code{validationStats} slot on an \code{S4} object}
\usage{
validationStats(object, ...) <- value
}
\arguments{
\item{object}{An \code{S4} object.}

\item{...}{Allow definition of additional parameters to this generic.}

\item{value}{A \code{data.frame} of validation statistics.}
}
\value{
None, updates the object
}
\description{
Setter for the \code{validationStats} slot on an \code{S4} object
}
\examples{
data(samplePCOSPmodel)
validationStats(samplePCOSPmodel) <- data.frame()

}
